% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPressure.R
\name{allPressure}
\alias{allPressure}
\title{Calculates all pressure indicators}
\usage{
allPressure(X, land, species.table, speciesinfo.table, landings.groups,
  FP.groups, cutoff, years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS". "YEAR" indicates the year the observation
was recorded, "ID" is an area code indicating where the observation was
recorded, "SPECIES" is a numeric code indicating the species sampled, and
"BIOMASS" is the corresponding biomass (stratified and corrected for
catchability as required).}

\item{land}{A dataframe of commercial landings data with columns "YEAR", "ID",
"SPECIES" and "CATCH". "YEAR" indicates the year the landing was recorded,
"ID" is an area code indicating where the landing was recorded, "SPECIES" is
a numeric code indicating the species landed, and "CATCH" is the
corresponding landed weight.}

\item{species.table}{If group does not equal "ALL", species.table is a table
with at least one column, where the column name is the string group, and the
column entries are the species codes indicating the species from X to
include in the calculation. species.table may also include columns for other
species groups; these will be ignored. If group = "ALL", this table is not
required. Default is species.table = NULL.}

\item{speciesinfo.table}{A table with columns "SPECIES" and the corresponding
"MAXLENGTH" and "TL" (maximum recorded length and trophic level). Entries in
the "SPECIES" column should be the unique values of species codes in X/X_length (or a
subset thereof). Other columns will be ignored.}

\item{landings.groups}{A vector indicating the species groups for which to
calculate the landings. Each entry must be a character string matching the
name of a column in species.groups. If landings.groups = NULL, no resource
potential indicators will be calculated.}

\item{FP.groups}{A dataframe with two columns, which must be named "group.X"
and "group.land". Each row holds the group names to calculate the fishing
pressure on a target group, with the numerator in column "group.land" and
the denominator in column "group.X". Each entry must be a character string
matching the name of a column in species.groups. If FP.groups = NULL, no
fishing pressure indicators will be calculated.}

\item{cutoff}{A vector containing minimum trophic level to include when
calculating the mean trophic level of the landings. Default is cutoff = c(0,
3.25), which will return the mean trophic level of the landings and the
marine trophic index.}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with columns "ID", "YEAR", and indicators
 corresponding to the arguments supplied to the function.
}
\description{
This function calculates all (or a subset) of the pressure
 indicators for \eqn{j} areas and \eqn{i} years.
}
\details{
This function calculates the pressure indicators: diversity of the
 target species, total landings, landings of target groups, fishing pressure
 on the community, fishing pressure on target groups, mean trophic level of
 the landings, and the marine trophic index. If data are not available to
 calculate one or more of these indicators, a subset will be returned. See
 the help file for the individual functions for information on how each
 indicator is calculated.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\seealso{
Other fishing pressure indicators: \code{\link{fishingPressure}},
  \code{\link{landings}}, \code{\link{meanTLLandings}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{fishing pressure indicators}
