% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassratio.R
\name{biomassratio}
\alias{biomassratio}
\title{Calculates the biomass (or abundance) ratio between two species groups}
\usage{
biomassratio(X, group1 = c("INVERTEBRATES", "PELAGIC"),
  group2 = c("GROUNDFISH"), metric = "BIOMASS", years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
designating where the observation was recorded. "SPECIES" is a numeric code
indicating the species sampled.}

\item{group1}{species group for the numerator (string), as defined in
functions "resourcePotential" and "speciesgroups"}

\item{group2}{species group for the denominator (string), as defined in
functions "resourcePotential" and "speciesgroups"}

\item{metric}{character string indicating which metric to use to calculate
indicator.}

\item{years}{vector of years for which to calculate indicator}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
  "group12group2".

  If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
  values is assigned NA.
}
\description{
This function calculates the biomass ration between two
  pre-defined species groups for \eqn{j} areas and \eqn{i} years.
}
\details{
Useful biomass ratio indicators include:
  \deqn{B_{invertebrates}/B_{demersal}} and \deqn{B_{pelagic}/B_{demersal}}

  **Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Bundy A, Heymans JJ, Morissette L, Savenkoff C (2009) Seals, cod and forage
  fish: A comparative exploration of variations in the theme of stock
  collapse and ecosystem change in four Northwest Atlantic ecosystems. Prog
  Oceanogr 81:188 206
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTrophicLevelCommunity}},
  \code{\link{predatoryFish}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{ecosystem structure and function indicators}
