% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassRatioInv2Dem.R
\name{biomassRatioInv2Dem}
\alias{biomassRatioInv2Dem}
\title{Calculates the invertebrate to demersal biomass ratio}
\usage{
biomassRatioInv2Dem(X, syear = yrs[i])
}
\arguments{
\item{X}{add text here}

\item{syear}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
  the invertebrate to demersal biomass ratio
}
\details{
**Recommended data: Fishery independent surveys, fish and
  invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Bundy A, Heymans JJ, Morissette L, Savenkoff C (2009) Seals, cod and forage
  fish: A comparative exploration of variations in the theme of stock
  collapse and ecosystem change in four Northwest Atlantic ecosystems. Prog
  Oceanogr 81:188–206
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassRatioPel2Dem}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTrophicLevelCommunityLength}},
  \code{\link{meanTrophicLevelCommunity}},
  \code{\link{predatoryFish}}, \code{\link{ratioIndicator}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{ecosystem structure and function indicators}
