% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityCondition.R
\name{communityCondition}
\alias{communityCondition}
\title{Calculates Fulton's Condition Index for the community (reference)}
\usage{
communityCondition(X, group, species.table = NULL, LenWt.table, years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", "LENGTH", and "ABUNDANCE". "YEAR" indicates the year the
observation was recorded, "ID" is an area code indicating where the
observation was recorded, and "SPECIES" is a numeric code indicating the
species sampled. "LENGTH" is the length class (cm) and "ABUNDANCE" is the
corresponding abundance at length (stratified and corrected for
catchability as required). Species for which there are no length data
should be assigned LENGTH = -99. These observations are removed by the
function.}

\item{group}{A character string indicating which species to include in the
indicator calculation. If group = "ALL", all species will be included;
otherwise, group should match a column name in species.table.}

\item{species.table}{A table with default species.table = NULL. If group =
"ALL", this table is not required. If group does not equal "ALL",
species.table is a table with at least one column, where the column name is
the same as group, and the column  entries are the species codes indicating
the species to be included in the calculation. species.table may also
include columns for different species groups; these will be ignored.}

\item{LenWt.table}{A table of annual length at weight data with 5 columns.
"YEAR", "ID", "SPECIES" correspond with those columns in X. "LENGTH" is
fish length at the corresponding "WEIGHT" (fish weight).}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
  "CommunityCondition_group".

  If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
  value is assigned NA.
}
\description{
This function calculates Fulton's condition index for \eqn{j}
  areas and \eqn{i} years.
}
\details{
Fulton's Condition Index (\eqn{K}): \deqn{K = \Sigma(K_j *
  A_j)/\Sigma A_j} where the sum is over all species, \eqn{j}, \eqn{A_j} is
  the abundance of species \eqn{j}, and \deqn{K_j = 100*W_j/L_j^3} where
  \eqn{W_j} is the mean weight at length \eqn{L} for species \eqn{j}.

  Recommended data: Fishery independent survey data or model output; fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  DFO. 2003. State of the Eastern Scotian Shelf ecosystem. Dartmouth, Nova
  Scotia

  Choi JS, Frank KT, Petrie BD, Leggett WC. 2005. Integrated Assessment of a
  Large Marine Ecosystem: a case study of the devolution of the Eastern
  Scotian Shelf, Canada. Oceanogr Mar Biol An Annu Rev 43:47–67
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassRatio}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTLCommunity}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
  Catalina Gomez, Alida Bundy
}
\concept{ecosystem structure and function indicators}
