% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityFultonK.R
\name{communityFultonK}
\alias{communityFultonK}
\title{Calculates Fulton's Condition Index for the community}
\usage{
communityFultonK(X, metric = "ABUNDANCE", gp = groups, yr = yrs[i],
  user.defined = F, group = c("FINFISH", "SKATES", "CLUPEIDS",
  "GROUNDFISH", "FLATFISH", "GADOIDS", "FORAGE", "LBENTHIVORE",
  "MBENTHIVORE", "PISCIVORE", "PLANKTIVORE", "ZOOPISCIVORE"), path = NA)
}
\arguments{
\item{X}{add text here}

\item{metric}{add text here}

\item{gp}{add text here}

\item{yr}{add text here}

\item{user.defined}{add text here}

\item{group}{add text here}

\item{path}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
  Fulton's Condition Index (K) for the community (weighted by abundance)
}
\details{
Fulton's Condition Index (K): \deqn{K = \Sigma(K_j * A_j)/\Sigma(A_j)}
  where \eqn{A_j} is the abundance of species j, and \deqn{K_j = W_j/L_j^3 x
  100} where \eqn{W_j} is the mean weight at length (L) for species j
  
  **Recommended data: Fishery independent surveys, fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  DFO (2003) State of the Eastern Scotian Shelf ecosystem. Dartmouth, Nova
  Scotia

  Choi JS, Frank KT, Petrie BD, Leggett WC (2005) Integrated Assessment of a
  Large Marine Ecosystem: a case study of the devolution of the Eastern
  Scotian Shelf, Canada. Oceanogr Mar Biol An Annu Rev 43:47–67

  Rochet M, Rice JC (2005) Do explicit criteria help in selecting indicators
  for ecosystem-based fisheries management? ICES J Mar Sci 62:528–539
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassRatioInv2Dem}},
  \code{\link{biomassRatioPel2Dem}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTrophicLevelCommunityLength}},
  \code{\link{meanTrophicLevelCommunity}},
  \code{\link{predatoryFish}}, \code{\link{ratioIndicator}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{ecosystem structure and function indicators}
