% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heips.R
\name{heips}
\alias{heips}
\title{Calculates Heip's Evenness Index}
\usage{
heips(X, group = c("FINFISH", "ALL"), metric = c("BIOMASS",
  "ABUNDANCE"))
}
\arguments{
\item{X}{add text here}

\item{group}{add text here}

\item{metric}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
  Heip's Evenness Index (HE) of the community
}
\details{
Heip's Evenness Index (HE): \deqn{HE = (exp^(H') - 1)/(S - 1)}
  where \eqn{H'} is the Shannon Index of Diversity and S is species richness.
  This index ranges from 0 to 1 and measures how equally the species richness
  contributes to the total abundance **or biomass?? of the community.

  **Recommended data: Fishery independent surveys, fish and invertebrates (NOT in tech report).
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Kenchington T, Kenchington E (2013) Biodiversity Metrics for Use in the
  Ecosystem Approach to Oceans Management. Can. Tech. Rep. Fish. Aquat. Sci.
  3059: vi+188p.
}
\seealso{
Other biodiversity indicators: \code{\link{hillN1}},
  \code{\link{hillN2}}, \code{\link{kemptonQ}},
  \code{\link{margalefSpeciesRichness}},
  \code{\link{margalef}},
  \code{\link{pielouSpeciesEvenness}},
  \code{\link{shannon}}, \code{\link{speciesrichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{biodiversity indicators}
