% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillN2.R
\name{hillN2}
\alias{hillN2}
\title{Calculates Hill's Species Dominance}
\usage{
hillN2(X, group = c("FINFISH", "ALL"), metric = c("BIOMASS",
  "ABUNDANCE"))
}
\arguments{
\item{X}{add text here}

\item{group}{add text here}

\item{metric}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
  Hill's Species Dominance (HillN2)
}
\details{
Hill's Species Dominance (HillN2): \deqn{HillN2 = 1/\Sigma(p_i^2)}
  \eqn{p_i} is the proportion of the total sample contributed by the i(th)
  species. HillN2 is the inverse of the Simpson's index. This index is
  sensitive to the evenness of the distribution of individuals between
  species.

  Data used: Fishery independent surveys, fish and invertebrates
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Greenstreet SP, Rogers SI (2006) Indicators of the health of the North Sea
  fish community: identifying reference levels for an ecosystem approach to
  management. ICES J Mar Sci J du Cons 63:573–593
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{kemptonQ}},
  \code{\link{margalefSpeciesRichness}},
  \code{\link{margalef}},
  \code{\link{pielouSpeciesEvenness}},
  \code{\link{shannon}}, \code{\link{speciesrichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{biodiversity indicators}
