% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseCVBiomass.R
\name{invCVBiomass}
\alias{invCVBiomass}
\title{Calculates the inverse of the coefficient of variation of the biomass}
\usage{
invCVBiomass(X, window = 5)
}
\arguments{
\item{X}{add text here}

\item{window}{Say that default is 5 for 5-year moving average. add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
  the inverse of the coefficient of variation of the biomass
  (InverseCVBiomass)
}
\details{
The inverse of the coefficient of variation of the biomass
  (InverseCVBiomass) was estimated as a five year moving average:
  \deqn{InverseCVBiomass = 1/(1/n)\Sigma(SD_i/Xbar_i)} The sum is from year
  \eqn{i} to year \eqn{i+n-1}, where \eqn{n=5} for the 5-year moving average.
  \eqn{SD_i} is the standard deviation of the mean biomass and \eqn{X_i} is the mean
  biomass for year i.

  Data used: Fishery independent surveys, fish and invertebrates
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Blanchard and Boucher 2001 (not in Tech refs)

  Shin YJ, Shannon LJ (2010) Using indicators for evaluating, comparing, and
  communicating the ecological status of exploited marine ecosystems. 1. the
  indiSeas project. ICES J Mar Sci 67:686–691
}
\seealso{
Other stability and resistance indicators: \code{\link{IVILandings}},
  \code{\link{biomassPerTL}}, \code{\link{meanMaxAge}},
  \code{\link{meanMaxL}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{stability and resistance indicators}
