% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InverseFishingPressure.r
\name{InverseFishingPressure}
\alias{InverseFishingPressure}
\title{Calculates inverse fishing pressure}
\usage{
InverseFishingPressure(path, land = dat, group, groups, qadj = T,
  qadjPostStrat = qadjPostStrat)
}
\arguments{
\item{path}{add text here}

\item{land}{add text here}

\item{group}{add text here}

\item{groups}{add text here}

\item{qadj}{add text here}

\item{qadjPostStrat}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
  inverse fishing pressure on a fished group (***not in Table S4)
}
\details{
Fishing pressure (InverseFP): \deqn{FP = B_FG/Y_FG} where \eqn{B_FG} is the
  biomass of the fished group(s) and \eqn{Y_FG} is the landed catch of the
  group(s). **Give examples of groups that can be used in function

  Recommended data: B: fishery independent surveys, Y: commercial fisheries
  landings
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Shin YJ, Bundy A, Shannon LJ, Simier M, Coll M, Fulton EA, Link JS, Jouffre
  D, Ojaveer H, MacKinson S, Heymans JJ, Raid T (2010) Can simple be useful
  and reliable? Using ecological indicators to represent and compare the
  states of marine ecosystems. ICES J Mar Sci 67:717–731
}
\seealso{
Other fishing pressure indicators: \code{\link{FishingPressure}},
  \code{\link{LandByGroup}},
  \code{\link{LandingsDiversity}},
  \code{\link{MeanTLLandings}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fishing pressure indicators}
