% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kemptonQ.R
\name{kemptonQ}
\alias{kemptonQ}
\title{Calculates Kempton's Biodiversity Index}
\usage{
kemptonQ(X, percentiles = c(0.25, 0.75), minTL = 3,
  metric = c("BIOMASS", "ABUNDANCE"))
}
\arguments{
\item{X}{add text here}

\item{percentiles}{R1 and R2 are defaulted to .25 and .75 add text here}

\item{metric}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
  Kempton's Biodiversity Index (Q)
}
\details{
Kempton's Biodiversity Index (Q): \deqn{Q = S/2*log(R_2/R_1)} S is
  the total number of species or functional groups, R1 and R2 are the lower
  and upper quartiles of the species abundance distribution. Q is a relative
  index of biomass diversity calculated from the Kempton’s Q75 index
  developed for expressing species diversity. This index includes those
  species or functional groups with a trophic level (TL) of three or higher.
  R1 and R2 are defaulted to .25 and .75 **Recommended data: Fishery
  independent surveys, fish and invertebrates
}
\references{
Ainsworth et al. 2006 (not in tech report)

  Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the selection and
  evaluation of ecological indicators. Can. Tech. Rep. Fish. Aquat. Sci.
  3232: xii + 212 p.

  Kempton R, Taylor L (1976) Models and statistics for species diversity.
  Nature 262:818-820
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{hillN2}},
  \code{\link{margalefSpeciesRichness}},
  \code{\link{margalef}},
  \code{\link{pielouSpeciesEvenness}},
  \code{\link{shannon}}, \code{\link{speciesrichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{biodiversity indicators}
