% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandByGroup.R
\name{LandByGroup}
\alias{LandByGroup}
\title{Calculates the landings per fished group}
\usage{
LandByGroup(land = dat, group = c("FINFISH", "CLUPEIDS", "GROUNDFISH",
  "FLATFISH", "GADOIDS", "FORAGE", "INVERTEBRATES", "ALL",
  "LARGE_PELAGIC"))
}
\arguments{
\item{land}{add text here}

\item{group}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
 the catch from commercial fisheries that is landed ashore.
}
\details{
**This text could change depending on structure of the data. Includes
 nine fished groups: Finfish, Skates, Clupeids, Flatfish, Gadoid, Groundfish,
 Invertebrates, Forage fish and Large pelagic fish.

 Recommended data: commercial fisheries landings
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Bundy A, Heymans JJ, Morissette L, Savenkoff C (2009) Seals, cod and forage
 fish: A comparative exploration of variations in the theme of stock collapse
 and ecosystem change in four Northwest Atlantic ecosystems. Prog Oceanogr
 81:188–206
}
\seealso{
Other fishing pressure indicators: \code{\link{FishingPressure}},
  \code{\link{InverseFishingPressure}},
  \code{\link{LandingsDiversity}},
  \code{\link{MeanTLLandings}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fishing pressure indicators}
