% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandingsDiversity.R
\name{LandingsDiversity}
\alias{LandingsDiversity}
\title{Calculates the diversity of the target species}
\usage{
LandingsDiversity(land = dat, grps = groups)
}
\arguments{
\item{land}{add text here}

\item{grps}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
 the diversity of the target species (\eqn{TS_y})
}
\details{
The diversity of the target species for year y (\eqn{TS_y}) is the
 count of the number of target species recorded in all trawl catches
 collected in that year.

 Recommended data: commercial fisheries landings, fish and invertebrates
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Degnbol P (2005) Indicators as a means of communicating knowledge. In: ICES
 Journal of Marine Science. p 606–611
}
\seealso{
Other fishing pressure indicators: \code{\link{FishingPressure}},
  \code{\link{InverseFishingPressure}},
  \code{\link{LandByGroup}}, \code{\link{MeanTLLandings}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fishing pressure indicators}
