% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeSpeciesIndicator.R
\name{largeSpeciesIndicator}
\alias{largeSpeciesIndicator}
\title{Calculates the Large Species Indicator}
\usage{
largeSpeciesIndicator(X, lmax = 85,
  linf.data.table = "indiseas_max_length", metric = "BIOMASS")
}
\arguments{
\item{X}{add text here}

\item{lmax}{(set to 85)}

\item{linf.data.table}{to delete}

\item{metric}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
 the Large Species Indicator (LSI)
}
\details{
Large Species Indicator (LSI): \deqn{LSI = \Sigma B_i(L_max >85
 cm)/\Sigma B_i} where \eqn{B_i} is biomass of individual species, i, and
 \eqn{L_max} is the maximum asymptotic length (cm; here the default is 85
 cm).

 Recommended data: Fishery independent surveys, fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shephard S, Fung T, Houle JE, Farnsworth KD, Reid DG, Rossberg AG (2012)
 Size-selective fishing drives species composition in the Celtic Sea. ICES J
 Mar Sci 69:223-234 

 Houle JE, Farnsworth KD, Rossberg AG, Reid DG (2012) Assessing the
 sensitivity and specificity of fish community indicators to management
 action. Can J Fish Aquat Sci 69:1065-1079
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassRatioInv2Dem}},
  \code{\link{biomassRatioPel2Dem}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{meanTrophicLevelCommunityLength}},
  \code{\link{meanTrophicLevelCommunity}},
  \code{\link{predatoryFish}}, \code{\link{ratioIndicator}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{ecosystem structure and function indicators}
