% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Length_Weight.R
\docType{data}
\name{Length_Weight}
\alias{Length_Weight}
\title{Example length-weight data}
\format{
This example data is a dataframe with 98 436 rows and 5 variables:
  \describe{ \item{YEAR}{Year the observation was recorded} \item{ID}{Area
  where the observation was recorded} \item{SPECIES}{Numeric code indicating
  the species sampled} \item{LENGTH}{Length (class)1 cm intervals)}
  \item{WEIGHT}{Corresponding weight of sampled fish} }
}
\source{
This is example data was modified from Fisheries and Oceans Canada
  summer research vessel trawl survey for the Scotian Shelf.
}
\usage{
Length_Weight
}
\description{
A dataframe of example length-weight data for two areas (AREA1
and AREA2) from 2014 - 2019.
}
\details{
Indicator calculated using length-weight data (among other arguments):
Community Condition
}
\keyword{datasets}
