% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margalefSpeciesRichness.R
\name{margalefSpeciesRichness}
\alias{margalefSpeciesRichness}
\title{Also calculates Margalef's species richness (see Margalef.R)}
\usage{
margalefSpeciesRichness(X, group = c("FINFISH", "ALL"),
  metric = c("BIOMASS", "ABUNDANCE"))
}
\arguments{
\item{X}{is probably a dataframe with certain columns.}

\item{group}{is where you select which groups to include}

\item{metric}{is where you choose if you want to calculate using biomass or
abundance}
}
\description{
This function takes a dataframe with columns **** and calculates
  Margalef's species richness
}
\details{
Margalef's species richness: \deqn{S = (S_y - 1)/log(F_y)} S is the
  count of the number of species recorded in all trawl catches collected in
  any one year (y). F is the total count of all individuals caught in all
  trawl catches in any one year (y). 
  
  **Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Greenstreet SP, Fraser HM, Rogers SI, Trenkel VM, Simpson SD, Pinnegar JK
  (2012) Redundancy in metrics describing the composition, structure, and
  functioning of the North Sea demersal fish community. ICES J Mar Sci
  69:8–22
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{hillN2}},
  \code{\link{kemptonQ}}, \code{\link{margalef}},
  \code{\link{pielouSpeciesEvenness}},
  \code{\link{shannon}}, \code{\link{speciesrichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{biodiversity indicators}
