% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanLengthCommunity.R
\name{meanLengthCommunity}
\alias{meanLengthCommunity}
\title{Calculates the mean length of the community weighted by biomass or
 abundance (Shin et al., 2010)}
\usage{
meanLengthCommunity(X_length, metric, years)
}
\arguments{
\item{X_length}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", "LENGTH", and "BIOMASS" and/or "ABUNDANCE". "YEAR"
indicates the year the observation was recorded, "ID" is an area code
indicating where the observation was recorded, and "SPECIES" is a numeric
code indicating the species sampled. "LENGTH" is the length class (cm) and
"BIOMASS" and "ABUNDANCE" are the corresponding biomass and abundance at
length (stratified and corrected for catchability as required). Species for
which there are no length data should be assigned LENGTH = -99. These
observations are removed by the function.}

\item{metric}{A character string indicating which column in X to use to
calculate indicator. Default is "BIOMASS".}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
 "MeanLength_metric".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 value is assigned NA.
}
\description{
This function calculates the mean length of the community
 weighted by biomass or abundance for \eqn{j} areas and \eqn{i} years.
}
\details{
\deqn{MeanLength = \Sigma(Length_m * metric_i)/\Sigma metric_j} where
 \eqn{Length_m} is the length (cm) of an individual in size class \eqn{m},
 \eqn{metric_i} is the biomass or abundance of species \eqn{i} and
 \eqn{metric_j} is the total biomass or abundance of the community.

 Recommended data: Fishery independent survey data or model output; fish and
 invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin, YJ, Shannon LJ, Bundy A, Coll M, Aydin K, Bez N, Blanchard JL, Borges,
 MF, Diallo I, Diaz E, Heymans JJ, Hill L, Johannesen E, Jouffre D, Kifani S,
 Labrosse P, Link JS, Mackinson S, Masski H, Möllmann C, Neira S, Ojaveer H,
 Abdallahi KM, Perry I, Thiao D, Yemane D, and Cury PM. 2010. Using
 indicators for evaluating, comparing and communicating the ecological status
 of exploited marine ecosystems. Part 2: Setting the scene. ICES Journal of
 Marine Science, 67: 692-716
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{structure and functioning indicators}
