% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMaxAge.R
\name{meanMaxAge}
\alias{meanMaxAge}
\title{Calculates the mean maximum age of fish in the community}
\usage{
meanMaxAge(X, table.of.age.data = "INDISEAS_MAX_AGE",
  metric = c("BIOMASS", "ABUNDANCE"))
}
\arguments{
\item{X}{add text here}

\item{table.of.age.data}{add text here --or delete}

\item{metric}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
 the mean maximum age (MMA) of fish in the community
}
\details{
Mean Maximum Lifespan (MMA): \deqn{MMA = \Sigma
 (age_max,i*B_i)/\Sigma B_i} where \eqn{B_i} is biomass of species i. The
 mean lifespan or longevity is considered to be a fixed parameter per
 species. Lifespan may vary under fishing pressure, so IndiSeas adopted the
 maximum longevity observed for each species. The variation of this indicator
 captures changes in species composition. **Check that i is subscript in PDF

 Recommended data: Fishery independent surveys, fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin YJ, Bundy A, Shannon LJ, Simier M, Coll M, Fulton EA, Link JS, Jouffre
 D, Ojaveer H, MacKinson S, Heymans JJ, Raid T (2010) Can simple be useful
 and reliable? Using ecological indicators to represent and compare the
 states of marine ecosystems. ICES J Mar Sci 67:717-731
}
\seealso{
Other stability and resistance indicators: \code{\link{IVILandings}},
  \code{\link{biomassPerTL}}, \code{\link{invCVBiomass}},
  \code{\link{meanMaxL}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{stability and resistance indicators}
