% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMaxL.R
\name{meanMaxL}
\alias{meanMaxL}
\title{Calculates the mean maximum length of fish in the community}
\usage{
meanMaxL(X, lmax.table, metric = c("BIOMASS", "ABUNDANCE"), years)
}
\arguments{
\item{X}{X dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", "FLEN", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area
code designating where the observation was recorded (a string). "SPECIES" is
a numeric code indicating the species sampled. "FLEN" is the length class
(cm) and "BIOMASS" and "ABUNDANCE" are the corresponding biomass and
abundance at length. Species for which there are no length data should be
assigned FLEN = -99. These observations are removed by the function.}

\item{lmax.table}{table with 2 columns: "SPECIES" and "MAXLEN99", the maximum
recorded length of the corresponding species.}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate indicator.}

\item{years}{vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
 "MMlength_metric".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 value is assigned NA.
}
\description{
This function calculates the mean maximum length (MML) of fish in
 the community weighted by biomass or abundance for \eqn{j} areas and \eqn{i}
 years.
}
\details{
Mean Maximum Length (MML): \deqn{MML = \Sigma (L_{max,i}*M_i)/\Sigma
 M_i} where \eqn{L_{max,i}} is the maximum asymptotic length (cm) of species
 \eqn{i}, and \eqn{M_i} is biomass or abundance of species \eqn{i} (excluding
 invertebrates).

 Recommended data: Fishery independent surveys, fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Houle JE, Farnsworth KD, Rossberg AG, Reid DG (2012) Assessing the
 sensitivity and specificity of fish community indicators to management
 action. Can J Fish Aquat Sci 69:1065-1079

 Shin YJ, Rochet MJ, Jennings S, Field JG, Gislason H (2005) Using size-based
 indicators to evaluate the ecosystem effects of fishing. In: ICES Journal of
 Marine Science. p 384-396
}
\seealso{
Other stability and resistance indicators: \code{\link{IVILandings}},
  \code{\link{biomassPerTL}}, \code{\link{invCVBiomass}},
  \code{\link{meanMaxAge}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{stability and resistance indicators}
