% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMaxLength.R
\name{meanMaxLength}
\alias{meanMaxLength}
\title{Calculates the mean maximum length of fish in the community (Shin et
 al., 2005)}
\usage{
meanMaxLength(X_length, group, species.table = NULL, lmax.table, metric,
  years)
}
\arguments{
\item{X_length}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", "LENGTH", and "BIOMASS" and/or "ABUNDANCE". "YEAR"
indicates the year the observation was recorded, "ID" is an area code
indicating where the observation was recorded, and "SPECIES" is a numeric
code indicating the species sampled. "LENGTH" is the length class (cm) and
"BIOMASS" and "ABUNDANCE" are the corresponding biomass and abundance at
length (stratified and corrected for catchability as required). Species for
which there are no length data should be assigned LENGTH = -99. These
observations are removed by the function.}

\item{group}{A character string indicating which species to include in the
indicator calculation. If group = "ALL", all species will be included;
otherwise, group should match a column name in species.table.}

\item{species.table}{If group does not equal "ALL", species.table is a table
with at least one column, where the column name is the string group, and the
column entries are the species codes indicating the species from X to
include in the calculation. species.table may also include columns for other
species groups; these will be ignored. If group = "ALL", this table is not
required. Default is species.table = NULL.}

\item{lmax.table}{A dataframe with columns "SPECIES" and "MAXLENGTH", the
maximum recorded length of the corresponding species. Entries in the
"SPECIES" column should be the unique values of species codes in X (or a
subset thereof). Other columns in lmax.table are ignored.}

\item{metric}{A character string indicating which column in X to use to
calculate indicator. Default is "BIOMASS".}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
 "MMLength_metric".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 value is assigned NA.
}
\description{
This function calculates the mean maximum length of fish in
 the community weighted by biomass or abundance for \eqn{j} areas and \eqn{i}
 years.
}
\details{
\deqn{Mean Maximum Length = \Sigma (L_{max,i}*M_i)/\Sigma
 M_i} where \eqn{L_{max,i}} is the maximum asymptotic length (cm) of species
 \eqn{i}, and \eqn{M_i} is biomass or abundance of species \eqn{i} (excluding
 invertebrates).

 Recommended data: Fishery independent survey data or model output; fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin YJ, Rochet MJ, Jennings S, Field JG, Gislason H. 2005. Using size-based
 indicators to evaluate the ecosystem effects of fishing. In: ICES Journal of
 Marine Science. p 384-396
}
\seealso{
Other stability and resistance indicators: \code{\link{CVBiomass}},
  \code{\link{IVILandings}}, \code{\link{allStability}},
  \code{\link{biomassPerTL}}, \code{\link{meanMaxAge}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{stability and resistance indicators}
