% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pielouSpeciesEvenness.R
\name{pielouSpeciesEvenness}
\alias{pielouSpeciesEvenness}
\title{Calculates Pielou's Species Evenness}
\usage{
pielouSpeciesEvenness(X, group = c("FINFISH", "ALL"),
  metric = c("BIOMASS", "ABUNDANCE"))
}
\arguments{
\item{X}{is probably a dataframe with certain columns.}

\item{group}{is where you select which groups to include}

\item{metric}{is where you choose if you want to calculate using biomass or
abundance}
}
\description{
This function takes a dataframe with columns **** and calculates
  Pielou's Species Evenness.
}
\details{
Pielou's Species Evenness: \deqn{J' = -\Sigma(p_i ln( p_i ))/ln(S)}
  \eqn{p_i} is the proportion of the total sample contributed by the i(th)
  species and S is the number of species recorded in the sample. Pielou's
  index is the Shannon-Weiner index computed for the sample S and represent a
  measure of evenness of the community.

  **Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Greenstreet SP, Rogers SI (2006) Indicators of the health of the North Sea
  fish community: identifying reference levels for an ecosystem approach to
  management. ICES J Mar Sci J du Cons 63:573–593
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{hillN2}},
  \code{\link{kemptonQ}},
  \code{\link{margalefSpeciesRichness}},
  \code{\link{margalef}}, \code{\link{shannon}},
  \code{\link{speciesrichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{biodiversity indicators}
