% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predatoryFish.R
\name{predatoryFish}
\alias{predatoryFish}
\title{Calculates the proportion of predatory fish in the community (Shin et
 al., 2010)}
\usage{
predatoryFish(X, pred.species, metric = "BIOMASS", years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS". "YEAR" indicates the year the observation
was recorded, "ID" is an area code indicating where the observation was
recorded, "SPECIES" is a numeric code indicating the species sampled, and
"BIOMASS" is the corresponding biomass (stratified and corrected for
catchability as required).}

\item{pred.species}{A vector of species codes for predatory fish}

\item{metric}{A character string indicating which column in X to use to
calculate indicator. Default is "BIOMASS".}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns: "ID", YEAR", and
 "PropPredatoryFish".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 value is assigned NA.
}
\description{
This function calculates the proportion of predatory fish in the
 community for \eqn{j} areas and \eqn{i} years.
}
\details{
Proportion of predatory fish in the community is estimated by:
 \deqn{PropPred = Biomass Predatory Fish Surveyed/Total Biomass Surveyed}
 Predatory fish species are defined as all surveyed fish species that are not
 largely planktivorous (i.e. phytoplankton and zooplankton feeders should be
 excluded; Shin et al. 2010). A fish species is classified as predatory if it
 is piscivorous, or if it feeds on invertebrates that are larger than the
 macrozooplankton category (0.2 cm). Detritivores should not be classified as
 predatory fish.

 This indicator captures changes in the trophic structure and changes in the
 functional diversity of fish in the ecosystem.

 Recommended data: Fishery independent survey data or model output; fish and
 invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin, YJ, Shannon LJ, Bundy A, Coll M, Aydin K, Bez N, Blanchard JL, Borges,
 MF, Diallo I, Diaz E, Heymans JJ, Hill L, Johannesen E, Jouffre D, Kifani S,
 Labrosse P, Link JS, Mackinson S, Masski H, Möllmann C, Neira S, Ojaveer H,
 Ould Mohammed Abdallahi ., Perry I, Thiao D, Yemane D, and Cury PM. 2010.
 Using indicators for evaluating, comparing and communicating the ecological
 status of exploited marine ecosystems. Part 2: Setting the scene. ICES
 Journal of Marine Science, 67: 692-716
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassratio}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTLCommunity}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{ecosystem structure and function indicators}
