% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioIndicator.R
\name{ratioIndicator}
\alias{ratioIndicator}
\title{not entirely sure that this one does}
\usage{
ratioIndicator(X, group, metric, user.defined)
}
\arguments{
\item{X}{add text here}

\item{group}{text}

\item{metric}{add text here}

\item{use.defined}{add text here}
}
\description{
Text
}
\details{
Text
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassRatioInv2Dem}},
  \code{\link{biomassRatioPel2Dem}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTrophicLevelCommunityLength}},
  \code{\link{meanTrophicLevelCommunity}},
  \code{\link{predatoryFish}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{ecosystem structure and function indicators}
