% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.info.R
\docType{data}
\name{species.info}
\alias{species.info}
\title{Species information}
\format{A dataframe with 306 rows and 6 variables: \describe{
 \item{SPECIES}{Species codes} \item{TL}{Trophic level of species caught in
 fishery independent surveys} \item{MAXLENGTH}{Maximum recorded length of of
 species caught in fishery independent surveys} \item{MAXAGE}{Maximum
 recorded age of of species caught in fishery independent surveys}
 \item{IVI}{Vulnerability of commercial species} \item{TL_LAND}{Trophic level
 of commercial species} }}
\source{
This is example data using species that are of interest the Scotian
 Shelf.
}
\usage{
species.info
}
\description{
Additional species information required to calculate several indicators (see
vignette).
}
\keyword{datasets}
