% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.table.R
\docType{data}
\name{species.table}
\alias{species.table}
\title{Example input for argument species.table}
\format{Dimensions will vary depending on the number of species groups of
  interest.

  This example data is a dataframe with 7000 rows and 15 variables:
  \describe{ \item{column names}{Each column is named after a species group
  of interest} \item{column entries}{The column entries are the species codes
  of species included in the group} }}
\source{
This is example data using species groups that are of interest the
  Scotian Shelf.
}
\usage{
species.table
}
\description{
A dataframe of the species groups of interest and the species included in
each.
}
\keyword{datasets}
