% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesGroups.R
\name{speciesGroups}
\alias{speciesGroups}
\title{Assigns species to species groups}
\usage{
speciesGroups(X, species.table = NULL, group)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "ABUNDANCE". "YEAR" indicates the year the observation
was recorded, "ID" is an area code indicating where the observation was
recorded, "SPECIES" is a numeric code indicating the species sampled, and
"ABUNDANCE" is the corresponding abundance (stratified and corrected for
catchability as required).}

\item{species.table}{A table with default species.table = NULL. If group =
"ALL", this table is not required. If group does not equal "ALL",
species.table is a table with at least one column, where the column name is
the same as group, and the column  entries are the species codes indicating
the species to be included in the calculation. species.table may also
include columns for different species groups; these will be ignored.}

\item{group}{A character string indicating which species to include in the
indicator calculation. If group = "ALL", all species will be included;
otherwise, group should match a column name in species.table.}
}
\value{
text here
}
\description{
Text here
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
