% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesRichness.R
\name{speciesrichness}
\alias{speciesrichness}
\title{Calculates species richness (S_y)}
\usage{
speciesrichness(X, group = c("FINFISH", "ALL"), grps)
}
\arguments{
\item{X}{is probably a dataframe with certain columns.}

\item{group}{is where you select which groups to include}

\item{grps}{not sure what this is because it is NOT called inside the
function}
}
\description{
This function takes a dataframe with columns **** and calculates
  species richness.
}
\details{
S is the count of the number of species recorded in all trawl
  catches collected in any one year (y). 
  
  Recommended data: Fishery independent
  surveys, fish and invertebrates
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
  selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
  Aquat. Sci. 3232: xii + 212 p.

  Greenstreet SP, Fraser HM, Rogers SI, Trenkel VM, Simpson SD, Pinnegar JK
  (2012) Redundancy in metrics describing the composition, structure, and
  functioning of the North Sea demersal fish community. ICES J Mar Sci
  69:8–22
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{hillN2}},
  \code{\link{kemptonQ}},
  \code{\link{margalefSpeciesRichness}},
  \code{\link{margalef}},
  \code{\link{pielouSpeciesEvenness}},
  \code{\link{shannon}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{biodiversity indicators}
