% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\usage{
test(p = NULL)
}
\arguments{
\item{p}{default is \code{NULL}. This is a parameter.}
}
\description{
This function is a test.
}
\seealso{
Other general_use: \code{\link{test2}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
