% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2.R
\name{test2}
\alias{test2}
\title{test2}
\usage{
test2(p = Sys.Date())
}
\arguments{
\item{p}{default is \code{NULL}. This is a parameter.}
}
\description{
This function is a test.
}
\seealso{
Other general_use: \code{\link{test}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
