% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_length.R
\docType{data}
\name{X_length}
\alias{X_length}
\title{Example length-based fishery independent survey data}
\format{This example data is a dataframe with 14378 rows and 6 variables:
  \describe{ \item{YEAR}{Year the observation was recorded} \item{ID}{Area
  where the observation was recorded} \item{SPECIES}{Numeric code indicating
  the species sampled} \item{LENGTH}{Length class of sample species, in cm}
  \item{ABUNDANCE}{Abundance of the species sampled, stratified and corrected
  for catchability} \item{BIOMASS}{Biomass of the sampled species, stratified
  and corrected for catchability} }}
\source{
This is made-up example data modified from Fisheries and Oceans
  Canada summer research vessel trawl survey for the Scotian Shelf Bioregion.
}
\usage{
X_length
}
\description{
A dataframe of example length-based fishery independent survey data for two
areas (AREA1 and AREA2) from 2014 - 2019.
}
\details{
Indicators calculated using length-based fishery independent survey data
(among other arguments): Community Condition, Large Fish Indicator, Mean
Length.
}
\keyword{datasets}
