% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allBiodiversity.R
\name{allBiodiversity}
\alias{allBiodiversity}
\title{Calculates all Biodiversity indicators}
\usage{
allBiodiversity(X, metric = "ABUNDANCE", group = "ALL",
  species.table = NULL, TL.table, percentiles = c(0.25, 0.75),
  minTL = 0, years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data or model output with
columns \code{YEAR}, \code{ID}, \code{SPECIES}, and \code{ABUNDANCE}.
\code{YEAR} indicates the year the observation was recorded, \code{ID} is an
area code indicating where the observation was recorded, \code{SPECIES} is a
numeric code indicating the species sampled, and \code{ABUNDANCE} is the
corresponding abundance (stratified and corrected for catchability as
required).}

\item{metric}{A character string indicating which column in \code{X} to use to
calculate the indicators. Default is \code{metric = "ABUNDANCE"}.}

\item{group}{A character string indicating the species group for which to
calculate the indicators. If \code{group = "ALL"}, all species will be
included; otherwise, \code{group} should match a column name in
\code{species.table}.}

\item{species.table}{A table with at least one column, where the column name
is the string \code{group}, and the column entries are species codes from
\code{X} indicating which species are included that group.
\code{species.table} may also include columns for other species groups;
these will be ignored. If \code{group = "ALL"}, this table is not required.
Default is \code{species.table = NULL}.}

\item{TL.table}{A dataframe with columns \code{SPECIES} and the corresponding
\code{TL} (trophic level). Entries in the \code{SPECIES} column should be
the unique values of species codes in \code{X} (or a subset thereof). Other
columns in \code{TL.table} are ignored. If \code{TL.table = NULL}, Kempton's
Q will not be calculated.}

\item{percentiles}{The percentiles used to determine R1 and R2 for calculating
Kempton's Q. Default is \code{percentiles = c(0.25, 0.75)}.}

\item{minTL}{Minimum trophic level for species included to calculate Kempton's
Q. Default is \code{minTL = 0}.}

\item{years}{A vector of years for which to calculate indicators.}
}
\value{
Returns a dataframe with columns \code{ID}, \code{YEAR}, and
 indicators corresponding to the arguments supplied to the function.

 If there is no data for an indicator at spatial scale \eqn{j} in year
 \eqn{i}, indicator value is assigned \code{NA}.
}
\description{
This function calculates all (or a subset) of the Biodiversity
 indicators for \eqn{j} areas and \eqn{i} years.
}
\details{
This function calculates the Biodiversity indicators: Species
 Richness, Shannon's Index of Diversity, Margalef's Species Richness,
 Pielou's Species Evenness, Hill's N1, Hill's N2, Heip's Evenness Index, and
 Kempton's Q. If trophic level data are not available to calculate Kempton's
 Q, the remaining Biodiversity indicators will be returned. See the help file
 for the individual functions for information on how each indicator is
 calculated.
}
\examples{
data(X)
data(species.info)
data(species.groups)

allBiodiversity(X = X, metric = "ABUNDANCE", group = "ALL", TL.table = species.info, 
   percentiles = c(.25, 0.75), minTL = 0, years = c(2014:2019))
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{hillN2}},
  \code{\link{kemptonQ}}, \code{\link{margalef}},
  \code{\link{pielouEvenness}}, \code{\link{shannon}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{biodiversity indicators}
