% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allPotential.R
\name{allPotential}
\alias{allPotential}
\title{Calculates all resource potential indicators}
\usage{
allPotential(X, land, species.table, speciesinfo.table, resource.groups,
  cutoff = 0, TE = 0.1, base.start, base.end, years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS". "YEAR" indicates the year the observation
was recorded, "ID" is an area code indicating where the observation was
recorded, "SPECIES" is a numeric code indicating the species sampled, and
"BIOMASS" is the corresponding biomass (stratified and corrected for
catchability as required).}

\item{land}{A dataframe of commercial landings data with columns "YEAR", "ID",
"SPECIES" and "CATCH". "YEAR" indicates the year the landing was recorded,
"ID" is an area code indicating where the landing was recorded, "SPECIES" is
a numeric code indicating the species landed, and "CATCH" is the
corresponding landed weight.}

\item{species.table}{If group does not equal "ALL", species.table is a table
with at least one column, where the column name is the string group, and the
column entries are the species codes indicating the species from X to
include in the calculation. species.table may also include columns for other
species groups; these will be ignored. If group = "ALL", this table is not
required. Default is species.table = NULL.}

\item{speciesinfo.table}{A table with columns "SPECIES" and the corresponding
"MAXLENGTH" and "TL" (maximum recorded length and trophic level). Entries in
the "SPECIES" column should be the unique values of species codes in X/X_length (or a
subset thereof). Other columns will be ignored.}

\item{resource.groups}{A vector indicating the species groups for which to
calculate the resource potential. Each entry must be a character string
matching the name of a column in species.groups. If resource.groups = NULL,
no resource potential indicators will be calculated.}

\item{cutoff}{The minimum trophic level of species to include.}

\item{TE}{Trophic efficiency. Default is TE = 0.1, i.e., a trophic efficiency
of 10%.}

\item{base.start}{Year indicating the beginning of the baseline period. The
average landings and average mean trophic level of the landings over the
baseline period are used as baseline values to calculate FiB (see
Details). land must include data for the baseline period.}

\item{base.end}{Year indicating the end of the baseline period. The
average landings and average mean trophic level of the landings over the
baseline period are used as baseline values to calculate FiB (see
Details). land must include data for the baseline period.}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with columns "ID", "YEAR", and indicators
 corresponding to the arguments supplied to the function.
}
\description{
This function calculates all (or a subset) of the resource
 potential indicators for \eqn{j} areas and \eqn{i} years.
}
\details{
This function calculates the resource potential indicators: abundance
 and biomass of the community, resource potential of predefined species
 groups, and fishing-in-balance. If data are not available to calculate
 one or more of these indicators, a subset will be returned. See the help
 file for the individual functions for information on how each indicator is
 calculated.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{allStructure}},
  \code{\link{biomassRatio}},
  \code{\link{communityCondition}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTLCommunity}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{ecosystem structure and function indicators}
