% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allStructure.R
\name{allStructure}
\alias{allStructure}
\title{Calculates all structure and function indicators}
\usage{
allStructure(X, X_length, resource.groups, condition.groups, ratio.groups,
  species.table, speciesinfo.table, LenWt.table, lmax, years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS". "YEAR" indicates the year the observation
was recorded, "ID" is an area code indicating where the observation was
recorded, "SPECIES" is a numeric code indicating the species sampled, and
"BIOMASS" is the corresponding biomass (stratified and corrected for
catchability as required).}

\item{X_length}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", "LENGTH", and "ABUNDANCE". "YEAR" indicates the year the
observation was recorded, "ID" is an area code indicating where the
observation was recorded, and "SPECIES" is a numeric code indicating the
species sampled. "LENGTH" is the length class (cm) and "ABUNDANCE" is the
corresponding abundance at length (stratified and corrected for
catchability as required). Species for which there are no length data
should be assigned LENGTH = -99. These observations are removed by the
function.}

\item{resource.groups}{A vector indicating the species groups for which to
calculate the resource potential. Each entry must be a character string
matching the name of a column in species.groups. If resource.groups = NULL,
no resource potential indicators will be calculated.}

\item{condition.groups}{A vector indicating the species groups for which to
calculate Fulton's community condition factor. Each entry must be a
character string matching the name of a column in species.groups. If
condition.groups = NULL, no community condition indicators will be
calculated.}

\item{ratio.groups}{A dataframe with two columns, which must be named "group1"
and "group2". Each row holds the group names for one biomass ratio, with the
numerator in column "group1" and the denominator in column "group2".  Each
entry must be a character string matching the name of a column in
species.groups. If ratio.groups = NULL, no biomass ratio. indicators will be
calculated.}

\item{species.table}{A table where the column names match the entries in
resource.groups, ratio.groups, and/or condition.groups. Column entries are
the species codes indicating the species to be included in each group.
species.table may also include columns for other species groups; these will
be ignored.}

\item{speciesinfo.table}{A table with columns "SPECIES" and the corresponding
"MAXLENGTH" and "TL" (maximum recorded length and trophic level). Other
columns will be ignored.}

\item{LenWt.table}{A table of annual length at weight data with 5 columns.
"YEAR", "ID", "SPECIES" correspond with those columns in X. "LENGTH" is
fish length at the corresponding "WEIGHT" (fish weight).}

\item{lmax}{The threshold for large fish (cm). Default is 85 cm (i.e., large
species are those with MAXLENGTH >= 85 cm)}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with columns "ID", "YEAR", and indicators
 corresponding to the arguments supplied to the function.
}
\description{
This function calculates all (or a subset of) structure and
 functioning indicators for \eqn{j} areas and \eqn{i} years.
}
\details{
Text here?
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassRatio}},
  \code{\link{communityCondition}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTLCommunity}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{ecosystem structure and function indicators}
