% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassPerTL.R
\name{biomassPerTL}
\alias{biomassPerTL}
\title{Calculates the biomass (or abundace??) per trophic level}
\usage{
biomassPerTL(X, table.of.trophic.level.data = "INDISEAS_WSS_TL",
  metric = c("BIOMASS", "ABUNDANCE"), TL.grouping = 1, path = NA)
}
\arguments{
\item{X}{add text here}

\item{table.of.trophic.level.data}{add text here}

\item{metric}{add text here}

\item{TL.grouping}{add text here}

\item{path}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
  the biomass (or abundance) per discrete trophic level.
}
\details{
**Recommended data: Fishery independent surveys, fish and
  invertebrates.
}
\references{
Coll et al 2006 (not in references!)

  Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the selection and
  evaluation of ecological indicators. Can. Tech. Rep. Fish. Aquat. Sci.
  3232: xii + 212 p.
}
\seealso{
Other stability and resistance indicators: \code{\link{IVILandings}},
  \code{\link{invCVBiomass}}, \code{\link{meanMaxAge}},
  \code{\link{meanMaxL}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{stability and resistance indicators}
