% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishingPressure.R
\name{fishingPressure}
\alias{fishingPressure}
\title{Calculates fishing pressure}
\usage{
fishingPressure(X, land, group.X, group.land, species.table = NULL,
  years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS". "YEAR" indicates the year the observation
was recorded, "ID" is an area code indicating where the observation was
recorded, "SPECIES" is a numeric code indicating the species sampled, and
"BIOMASS" is the corresponding biomass (stratified and corrected for
catchability as required).}

\item{land}{A dataframe of commercial landings data with columns "YEAR", "ID",
"SPECIES" and "CATCH". "YEAR" indicates the year the landing
was recorded, "ID" is an area code indicating where the landing was
recorded, "SPECIES" is a numeric code indicating the species landed, and
"CATCH" is the corresponding landed weight.}

\item{group.X}{A character string indicating which species to include in the
denominator. Must match the name of a column in species.table.}

\item{group.land}{A character string indicating which species to include in the
numerator. Must match the name of a column in species.table.}

\item{species.table}{A table with default species.table = NULL. If group =
"ALL", this table is not required. If group does not equal "ALL",
species.table is a table with at least one column, where the column name is
the same as group, and the column  entries are the species codes indicating
the species to be included in the calculation. species.table may also
include columns for different species groups; these will be ignored.}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
returns a dataframe with three columns: "ID", "YEAR", and "FP_group".

 If biomass of group is NA and landings of group are zero, fishing pressure
 is set to zero. Otherwise, if biomass of group is NA, fishing pressure is
 set to NA.
}
\description{
This function calculates fishing pressure on a fished group for
 \eqn{j} areas and \eqn{i} years.
}
\details{
Fishing pressure (FP): \deqn{FP = Y_{FG}/B_{FG}} where \eqn{B_{FG}}
 is the biomass of the fished group(s) and \eqn{Y_{FG}} is the landed catch
 of the group(s).

 This indicator measures the level of exploitation or total fishing pressure
 at the ecosystem level. Change in this indicator can result from change in
 \eqn{B_{FG}}, \eqn{Y_{FG}} or both. If \eqn{B_{FG}} and \eqn{Y_{FG}} change
 in the same direction, exploitation rate may not change.

 Recommended data: \eqn{B_{FG}}: fishery independent survey data, \eqn{Y_{FG}}:
 commercial fisheries landings
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin, YJ, Shannon LJ, Bundy A, Coll M, Aydin K, Bez N, Blanchard JL, Borges,
 MF, Diallo I, Diaz E, Heymans JJ, Hill L, Johannesen E, Jouffre D, Kifani S,
 Labrosse P, Link JS, Mackinson S, Masski H, Möllmann C, Neira S, Ojaveer H,
 Ould Mohammed Abdallahi ., Perry I, Thiao D, Yemane D, and Cury PM. 2010.
 Using indicators for evaluating, comparing and communicating the ecological
 status of exploited marine ecosystems. Part 2: Setting the scene. ICES
 Journal of Marine Science, 67: 692-716
}
\seealso{
Other fishing pressure indicators: \code{\link{landings}},
  \code{\link{meanTLLandings}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca} Cataline Gomez
}
\concept{fishing pressure indicators}
