% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishingPressure.R
\name{fishingPressure}
\alias{fishingPressure}
\title{Calculates fishing pressure}
\usage{
fishingPressure(X, land, FP.groups, species.table = NULL, years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data or model output with
columns \code{YEAR}, \code{ID}, \code{SPECIES}, and \code{BIOMASS}.
\code{YEAR} indicates the year the observation was recorded, \code{ID} is an
area code indicating where the observation was recorded, \code{SPECIES} is a
numeric code indicating the species sampled, and \code{BIOMASS} is the
corresponding biomass (stratified and corrected for catchability as
required).}

\item{land}{A dataframe of commercial landings data with columns \code{YEAR},
\code{ID}, \code{SPECIES} and \code{CATCH}. \code{YEAR} indicates the year
the landing was recorded, \code{ID} is an area code indicating where the
landing was recorded, \code{SPECIES} is a numeric code indicating the
species landed, and \code{CATCH} is the corresponding landed weight.}

\item{FP.groups}{A dataframe with two columns, which must be named
\code{group.land} and \code{group.X}. Each row holds the group names to
calculate the fishing pressure on a target group, with the numerator in
column \code{group.land} and the denominator in column {group.X}.  Each
entry must be a character string matching the name of a column in
\code{species.groups}.}

\item{species.table}{A table with column names that match the entries of
\code{FP.groups}. The entries in each column are the species codes for the
species included in that group. Species codes should be a subset of those in
the \code{SPECIES} column of \code{land} (for \code{group.land}) or \code{X}
(for \code{group.X}). \code{species.table} may also include columns for
other species groups; these will be ignored. Note that an entry in
\code{FP.groups} could be \code{"ALL"}. In this case, a column in
\code{species.table} named \code{"ALL"} is not required; the function will
automatically include all species in \code{land} and/or \code{X}.}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with columns: \code{ID} and \code{YEAR}, and a
 column for each target group, named \code{FP_group.X}.

 If biomass of \code{group.X} is \code{NA} and landings of \code{group.land}
 are zero, fishing pressure is set to zero. Otherwise, if biomass of
 \code{group.X} is \code{NA}, fishing pressure is set to \code{NA}.
}
\description{
This function calculates fishing pressure on a fished group for
 \eqn{j} areas and \eqn{i} years.
}
\details{
Fishing pressure (FP): \deqn{FP = Y_{FG}/B_{FG}} where \eqn{B_{FG}}
 is the biomass of the fished group(s) and \eqn{Y_{FG}} is the landed catch
 of the group(s).

 This indicator measures the level of exploitation or total fishing pressure
 at the ecosystem level. Change in this indicator can result from change in
 \eqn{B_{FG}}, \eqn{Y_{FG}} or both. If \eqn{B_{FG}} and \eqn{Y_{FG}} change
 in the same direction, exploitation rate may not change.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin, YJ, Shannon LJ, Bundy A, Coll M, Aydin K, Bez N, Blanchard JL, Borges,
 MF, Diallo I, Diaz E, Heymans JJ, Hill L, Johannesen E, Jouffre D, Kifani S,
 Labrosse P, Link JS, Mackinson S, Masski H, Möllmann C, Neira S, Ojaveer H,
 Abdallahi KM, Perry I, Thiao D, Yemane D, and Cury PM. 2010. Using
 indicators for evaluating, comparing and communicating the ecological status
 of exploited marine ecosystems. Part 2: Setting the scene. ICES Journal of
 Marine Science, 67: 692-716
}
\seealso{
Other fishing pressure indicators: \code{\link{allPressure}},
  \code{\link{landings}}, \code{\link{meanTLLandings}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{fishing pressure indicators}
