% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillN1.R
\name{hillN1}
\alias{hillN1}
\title{Calculates Hill's Index of Diversity (N1)}
\usage{
hillN1(X, group, species.table = NULL, metric = "ABUNDANCE", years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data or model output with
columns \code{YEAR}, \code{ID}, \code{SPECIES}, and \code{ABUNDANCE}.
\code{YEAR} indicates the year the observation was recorded, \code{ID} is an
area code indicating where the observation was recorded, \code{SPECIES} is a
numeric code indicating the species sampled, and \code{ABUNDANCE} is the
corresponding abundance (stratified and corrected for catchability as
required).}

\item{group}{A character string indicating which species to include in the
indicator calculation. If \code{group = "ALL"}, all species will be
included; otherwise, \code{group} should match a column name in
\code{species.table}.}

\item{species.table}{A table with at least one column, where the column name
is the string \code{group}, and the column entries are species codes from
\code{X} indicating which species are included that group.
\code{species.table} may also include columns for other species groups;
these will be ignored. If \code{group = "ALL"}, this table is not required.
Default is \code{species.table = NULL}.}

\item{metric}{A character string indicating which column in \code{X} to use to
calculate the indicator. Default is \code{metric = "ABUNDANCE"}.}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns: \code{ID}, \code{YEAR}, and
 \code{HillDiversity}.

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 values is assigned \code{NA}.
}
\description{
This function calculates Hill's Index of Diversity (N1) for
 \eqn{j} areas and \eqn{i} years.
}
\details{
Hill's N1 Diversity Index is the exponential of the Shannon-Weiner
 index.\deqn{HillN1 = e^{-\Sigma p_i ln( p_i )}} \eqn{p_i} is the proportion
 of the total sample contributed by the i(th) species and \eqn{S} is the
 number of species recorded in the sample. This index is sensitive to the
 number of species recorded in the sample (Hill, 1973).
}
\examples{
data(X)
hillN1(X, group = "ALL", metric = "ABUNDANCE", years = c(2014:2019))
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Greenstreet SP, Rogers SI. 2006. Indicators of the health of the North Sea
 fish community: identifying reference levels for an ecosystem approach to
 management. ICES J Mar Sci J du Cons 63:573-593

 Hill MO. 1973. Diversity and evenness: a unifying notation and its
 consequences. Ecology 54: 427-431.
}
\seealso{
Other biodiversity indicators: \code{\link{allBiodiversity}},
  \code{\link{heips}}, \code{\link{hillN2}},
  \code{\link{kemptonQ}}, \code{\link{margalef}},
  \code{\link{pielouEvenness}}, \code{\link{shannon}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca}, Catalina
 Gomez, Alida Bundy
}
\concept{biodiversity indicators}
