% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillN2.R
\name{hillN2}
\alias{hillN2}
\title{Calculates Hill's Species Dominance}
\usage{
hillN2(X, group = c("FINFISH", "ALL"), metric = c("BIOMASS",
  "ABUNDANCE"), years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
designating where the observation was recorded. "SPECIES" is a numeric code
indicating the species sampled.}

\item{group}{character string indicating which species to include, either
"ALL" or "FINFISH". Note that this subsetting is based on the Fisheries and
Oceans Canada species codes for the Scotian Shelf. For other regions it may
be prudent to subsetdata to species groups of interest prior to using the
function and then choose group = "ALL". Type ?speciesgroups for more
information.}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate the indicator.}

\item{years}{vector of years for which to calculate indicator}
}
\value{
Returns a dataframe with 3 columns: "ID", YEAR", and "HillDominance"
}
\description{
This function takes a dataframe of fisheries independent survey
 data and calculates Hill's species dominance for \eqn{j} areas and \eqn{i}
 years.
}
\details{
Hill's Species Dominance (HillN2): \deqn{HillN2 = 1/\Sigma p_i^2}
 \eqn{p_i} is the proportion of the total sample contributed by the i(th)
 species. HillN2 is the inverse of the Simpson's index. This index is
 sensitive to the evenness of the distribution of individuals between
 species.

 Recommended data: Fishery independent surveys, fish and invertebrates
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Greenstreet SP, Rogers SI (2006) Indicators of the health of the North Sea
 fish community: identifying reference levels for an ecosystem approach to
 management. ICES J Mar Sci J du Cons 63:573-593
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{kemptonQ}},
  \code{\link{margalef}},
  \code{\link{pielouSpeciesEvenness}},
  \code{\link{shannon}}, \code{\link{speciesrichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
  \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{biodiversity indicators}
