% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillN2.R
\name{hillN2}
\alias{hillN2}
\title{Calculates Hill's species dominance (Hill, 1973)}
\usage{
hillN2(X, group, species.table = NULL, metric = "ABUNDANCE", years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "ABUNDANCE". "YEAR" indicates the year the observation
was recorded, "ID" is an area code indicating where the observation was
recorded, "SPECIES" is a numeric code indicating the species sampled, and
"ABUNDANCE" is the corresponding abundance (stratified and corrected for
catchability as required).}

\item{group}{A character string indicating which species to include in the
indicator calculation. If group = "ALL", all species will be included;
otherwise, group should match a column name in species.table.}

\item{species.table}{If group does not equal "ALL", species.table is a table
with at least one column, where the column name is the string group, and the
column entries are the species codes indicating the species from X to
include in the calculation. species.table may also include columns for other
species groups; these will be ignored. If group = "ALL", this table is not
required. Default is species.table = NULL.}

\item{metric}{A character string indicating which column in X to use to
calculate indicator. Default is "ABUNDANCE".}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns: "ID", YEAR", and "HillDominance".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 values is assigned NA.
}
\description{
This function calculates Hill's species dominance for \eqn{j}
 areas and \eqn{i} years.
}
\details{
Hill's Species Dominance (HillN2): \deqn{HillN2 = 1/\Sigma p_i^2}
 \eqn{p_i} is the proportion of the total sample contributed by the i(th)
 species. HillN2 is the inverse of the Simpson's index. This index is
 sensitive to the evenness of the distribution of individuals between
 species.

 Recommended data: Fishery independent surveys or model output; fish and
 invertebrates
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Greenstreet SP, Rogers SI. 2006. Indicators of the health of the North Sea
 fish community: identifying reference levels for an ecosystem approach to
 management. ICES J Mar Sci J du Cons 63:573-593

 Hill MO. 1973. Diversity and evenness: a unifying notation and its
 consequences. Ecology 54: 427-431.
}
\seealso{
Other biodiversity indicators: \code{\link{allBiodiversity}},
  \code{\link{heips}}, \code{\link{hillN1}},
  \code{\link{kemptonQ}}, \code{\link{margalef}},
  \code{\link{pielouEvenness}}, \code{\link{shannon}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{biodiversity indicators}
