% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVILandings.R
\name{IVILandings}
\alias{IVILandings}
\title{Calculates the intrinsic vulnerability index of the catch}
\usage{
IVILandings(land, IVI.table, negative = FALSE, years)
}
\arguments{
\item{land}{A dataframe of commercial landings data with columns "YEAR", "ID",
"SPECIES" and "CATCH". "YEAR" indicates the year the landing
was recorded, "ID" is an area code indicating where the landing was
recorded, "SPECIES" is a numeric code indicating the species landed, and
"CATCH" is the corresponding landed weight.}

\item{IVI.table}{dataframe with two columns: "SPECIES" and "IVI". "SPECIES" is
the fisheries-independent numeric species code, and "IVI" is the
corresponding intrinsic vulnerability index.}

\item{negative}{If negative = TRUE, the indicator will be multipled by -1 so
that the expected response is to decrease with increasing fishing pressure.
Default is negative = FALSE.}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
returns a dataframe with three columns: "ID", "YEAR", and
 "IVILandings".

 If there are no observations in land for spatial scale \eqn{j} in year
 \eqn{i}, indicator value is assigned NA.
}
\description{
This calculates the intrinsic vulnerability index (IVI) of the
 catch for \eqn{j} areas and \eqn{i} years.
}
\details{
The weighted intrinsic vulnerability index (IVI) of species in the
 catch was estimated as: \deqn{IVI = \Sigma(IVI_j * C_j)/\Sigma C_j} where
 the sum is over all species, j. The vulnerability for each species was
 determined by considering several traits including maximum length, age at
 first maturity, longevity, von Bertalanffy growth parameter, natural
 mortality, fecundity, spatial behaviour and geographic range (e.g. species
 with larger body size, higher longevity, higher age at maturity, and lower
 growth rates have higher vulnerability indexes and should be less able to
 sustain high fishing mortality; Cheung et al. 2005). The index values ranges
 from 1 to 100, with 100 being the most vulnerable. See Appendix 1 from
 Cheung et al., 2007:
 \url{http://www.int-res.com/articles/suppl/m333p001_app.pdf}

 Recommended data: Commercial fisheries landings, fish
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Cheung, William W. L., watson, Reg, Morato, Telmo, Pitcher, Tony J., Pauly,
 Daniel (2007) Intrinsic vulnerability in the global fish catch. Mar Ecol
 Prog Ser 333: 1 - 12
}
\seealso{
Other stability and resistance indicators: \code{\link{CVBiomass}},
  \code{\link{biomassPerTL}}, \code{\link{meanMaxAge}},
  \code{\link{meanMaxLength}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{stability and resistance indicators}
