% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kemptonQ.R
\name{kemptonQ}
\alias{kemptonQ}
\title{Calculates Kempton's Biodiversity Index}
\usage{
kemptonQ(X, TL.table, percentiles = c(0.25, 0.75), minTL = 3,
  group = "ALL", metric = c("BIOMASS", "ABUNDANCE"), years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
designating where the observation was recorded. "SPECIES" is a numeric code
indicating the species sampled.}

\item{TL.table}{table with columns "SPECIES" and "TL", where the "SPECIES"
codes match those in X, and "TL" is the corresponding trophic level.}

\item{percentiles}{percentiles used to determine R1 and R2. Default here is
percentiles = c(0.25, 0.75).}

\item{minTL}{minimum trophic level for species included in the calculation.
Default is minTL = 3.}

\item{group}{character string indicating which species to include, either
"ALL" or "FINFISH". Note that this subsetting is based on the Fisheries and
Oceans Canada species codes for the Scotian Shelf. For other regions it may
be prudent to subsetdata to species groups of interest prior to using the
function and then choose group = "ALL". Type ?speciesgroups for more
information.}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate the indicator.}

\item{years}{vector of years for which to calculate indicator}
}
\value{
Returns a dataframe with 3 columns: "ID", YEAR", and "KemptonQ".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 values is assigned NA.
}
\description{
This function takes a dataframe of fisheries independent survey
 data and calculates Kempton's Biodiversity index (Q) for \eqn{j} areas and
 \eqn{i} years.
}
\details{
Kempton's Biodiversity Index \eqn{(Q)} is a relative index of biomass
 diversity calculated from the Kempton's Q75 index developed for expressing
 species diversity. This index includes those species or functional groups
 with a trophic level of three or higher. \deqn{Q = S*(p_2 -
 p_1)*log(R_2/R_1)} \eqn{S} is the total number of species or functional
 groups, \eqn{p_1} and  \eqn{p_2} are the lower and upper percentiles of
 interest and \eqn{R_1} and \eqn{R_2} are the corresponding lower and upper
 quartiles of the species abundance distribution. \eqn{p_1} and \eqn{p_2} are
 defaulted to 0.25 and 0.75, respectively.

 **Recommended data: Fishery independent surveys, fish and invertebrates
}
\references{
(Alida - this ref was not in tech report. Could be this one or the
 one below) Ainsworth, C, Pitcher, T (2006) Modifying Kempton's species
 diversity index for use with ecosystem simulation models. Ecological
 Indicators. 6. 623-630. 10.1016/j.ecolind.2005.08.024.

 Ainsworth, C, Varkey, D, Pitcher, TJ  (2006)  Preliminary ecosystem
 simulation models for the Bird’s Head Seascape, Papua.  Mid-term narrative
 technical report.  Birds Head Seascape Ecosystem-Based Management Project.
 University of British Columbia Fisheries Centre.  December, 2006, 274 pp.

 Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the selection and
 evaluation of ecological indicators. Can. Tech. Rep. Fish. Aquat. Sci. 3232:
 xii + 212 p.

 Kempton R, Taylor L (1976) Models and statistics for species diversity.
 Nature 262:818-820
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{hillN2}},
  \code{\link{margalef}},
  \code{\link{pielouSpeciesEvenness}},
  \code{\link{shannon}}, \code{\link{speciesrichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{biodiversity indicators}
