% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandByGroup.R
\name{LandByGroup}
\alias{LandByGroup}
\title{Calculates the landings per fished group}
\usage{
LandByGroup(land, group = c("FINFISH", "CLUPEIDS", "GROUNDFISH",
  "FLATFISH", "GADOIDS", "FORAGE", "INVERTEBRATES", "ALL",
  "LARGE_PELAGIC"), years)
}
\arguments{
\item{land}{dataframe of commercial landings data with columns "YEAR", "ID",
"ALLCODES" and "CATCH". "ID" is an area code designating where the
observation was recorded. "ALLCODES" is a numeric commercial species code
indicating the species landed, and "CATCH" is the corresponding landed
weight. Additional columns are required for each species group of interest.
These columns have value of "1" in the rows of species included in the group
and "NA" in all other rows.}

\item{group}{string indicating the species group for which to calculate the
landings. Should match one of the column names of land. If group = "ALL" the
total landings of all species will be calculated.}

\item{years}{vector of years for which to calculate indicator}
}
\value{
returns a dataframe with three columns: "ID", "YEAR", and
 "group_Landings".

 If there is no data for a given year, the indicator value is set to 0.
}
\description{
This function calculates the the catch of predefined speceies
 groups from commercial fisheries that is landed ashore for \eqn{j} areas and
 \eqn{i} years.
}
\details{
Calculates the total landings of predefined species groups.

 Recommended data: commercial fisheries landings
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Bundy A, Heymans JJ, Morissette L, Savenkoff C (2009) Seals, cod and forage
 fish: A comparative exploration of variations in the theme of stock collapse
 and ecosystem change in four Northwest Atlantic ecosystems. Prog Oceanogr
 81:188–206
}
\seealso{
Other fishing pressure indicators: \code{\link{FishingPressure}},
  \code{\link{InverseFishingPressure}},
  \code{\link{MeanTLLandings}},
  \code{\link{speciesrichness}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{fishing pressure indicators}
