% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings.R
\name{landings}
\alias{landings}
\title{Calculates the landings per fished group}
\usage{
landings(land, group, species.table = NULL, years)
}
\arguments{
\item{land}{A dataframe of commercial landings data with columns "YEAR", "ID",
"SPECIES" and "CATCH". "YEAR" indicates the year the landing was recorded,
"ID" is an area code indicating where the landing was recorded, "SPECIES" is
a numeric code indicating the species landed, and "CATCH" is the
corresponding landed weight.}

\item{group}{A character string indicating which species to include in the
indicator calculation. If group = "ALL", all species will be included;
otherwise, group should match a column name in species.table.}

\item{species.table}{If group does not equal "ALL", species.table is a table
with at least one column, where the column name is the string group, and the
column entries are the species codes indicating the species from X to
include in the calculation. species.table may also include columns for other
species groups; these will be ignored. If group = "ALL", this table is not
required. Default is species.table = NULL.}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
returns a dataframe with three columns: "ID", "YEAR", and
 "group_Landings".

 If there is no data for a given year, the indicator value is set to 0.
}
\description{
This function calculates the commercial landings of predefined
 speceies groups for \eqn{j} areas and \eqn{i} years.
}
\details{
Calculates the total landings of predefined species groups.

 Recommended data: commercial fisheries landings
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\seealso{
Other fishing pressure indicators: \code{\link{allPressure}},
  \code{\link{fishingPressure}},
  \code{\link{meanTLLandings}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{fishing pressure indicators}
