% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeFishIndicator.R
\name{largeFishIndicator}
\alias{largeFishIndicator}
\title{Calculates the Large Fish Indicator}
\usage{
largeFishIndicator(X, metric, large.fish = 35, years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", "FLEN", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area
code designating where the observation was recorded (a string). "SPECIES" is
a numeric code indicating the species sampled. "FLEN" is the length class
(cm) and "BIOMASS" and "ABUNDANCE" are the corresponding biomass and
abundance at length. Species for which there are no length data should be
assigned FLEN = -99. These observations are removed by the function.}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate indicator.}

\item{large.fish}{threshold for large fish (cm). Default is 35 cm (i.e., large
fish are those with X$FLEN >= 35 cm)}

\item{years}{vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
 "LargeFishIndicator".

 If there are no observations of large fish or no observations in X for
 spatial scale \eqn{j} in year \eqn{i}, indicator value is assigned NA.
}
\description{
This function takes a dataframe of length-based fisheries
 independent survey data and calculates the Large Fish Indicator (LFI) for
 \eqn{j} areas and \eqn{i} years.
}
\details{
Large Fish Indicator (LFI): \deqn{LFI = \Sigma B_m(L >50 cm)/\Sigma
 B_m} \eqn{B_m} is biomass of individuals in a body size class centred at
 mass m, and \eqn{L} is the length (cm) of an individual. This indicator
 describes the proportion (by weight) of the fish community that is larger
 than some length threshold (default here is 35 cm, **check this: i.e., the
 proportion of biomass occupying the top predator trophic level).

 Recommended data: Fishery independent surveys, fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Greenstreet SP, Fraser HM, Rogers SI, Trenkel VM, Simpson SD, Pinnegar JK
 (2012) Redundancy in metrics describing the composition, structure, and
 functioning of the North Sea demersal fish community. ICES J Mar Sci 69:8-22

 Houle JE, Farnsworth KD, Rossberg AG, Reid DG (2012) Assessing the
 sensitivity and specificity of fish community indicators to management
 action. Can J Fish Aquat Sci 69:1065-1079
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassratio}},
  \code{\link{communityFultonK}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTrophicLevelCommunity}},
  \code{\link{predatoryFish}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{ecosystem structure and function indicators}
