% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeSpeciesIndicator.R
\name{largeSpeciesIndicator}
\alias{largeSpeciesIndicator}
\title{Calculates the Large Species Indicator}
\usage{
largeSpeciesIndicator(X, lmax.table, lmax = 85, metric = "BIOMASS",
  years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
designating where the observation was recorded. "SPECIES" is a numeric code
indicating the species sampled.}

\item{lmax.table}{table with 2 columns: "SPECIES" and "MAXLEN99", the maximum
recorded length of the corresponding species.}

\item{lmax}{threshold for large fish (cm). Default is 85 cm (i.e., large
species are those with MAXLEN99 >= 85 cm)}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate indicator.}

\item{years}{vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
 "LargeFishIndicator"
}
\description{
This function takes a dataframe of length-based fisheries
 independent survey data and calculates the Large Species Indicator (LSI) for
 \eqn{j} areas and \eqn{i} years.
}
\details{
Large Species Indicator (LSI): \deqn{LSI = \Sigma B_i(L_{max} >85
 cm)/\Sigma B_i} where \eqn{B_i} is biomass of individual species, \eqn{i},
 and \eqn{L_{max}} is the maximum asymptotic length (cm; here the default is
 85 cm).

 Recommended data: Fishery independent surveys, fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shephard S, Fung T, Houle JE, Farnsworth KD, Reid DG, Rossberg AG (2012)
 Size-selective fishing drives species composition in the Celtic Sea. ICES J
 Mar Sci 69:223-234

 Houle JE, Farnsworth KD, Rossberg AG, Reid DG (2012) Assessing the
 sensitivity and specificity of fish community indicators to management
 action. Can J Fish Aquat Sci 69:1065-1079
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassratio}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{meanTrophicLevelCommunity}},
  \code{\link{predatoryFish}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{ecosystem structure and function indicators}
