% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanLengthCommunity.R
\name{meanLengthCommunity}
\alias{meanLengthCommunity}
\title{Calculates the mean length of the community weighted by biomass or}
\usage{
meanLengthCommunity(X, metric = c("BIOMASS", "ABUNDANCE"), years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", "FLEN", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area
code designating where the observation was recorded (a string). "SPECIES" is
a numeric code indicating the species sampled. "FLEN" is the length class
(cm) and "BIOMASS" and "ABUNDANCE" are the corresponding biomass and
abundance at length. Species for which there are no length data should be
assigned FLEN = -99. These observations are removed by the function.}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate indicator.}

\item{years}{vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
 "MeanLength_metric".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 value is assigned NA.
}
\description{
This function calculates the mean length of the community
 weighted by biomass or abundance for \eqn{j} areas and \eqn{i} years.
}
\details{
\deqn{MeanLength = \Sigma(Length_m * metric_i)/\Sigma metric_j} where
 \eqn{Length_m} is the length (cm) of an individual in size class \eqn{m},
 \eqn{B_i} is the biomass or abundance of species \eqn{i} and \eqn{metric_j}
 is the total biomass or abundance of the community.

 **Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Houle JE, Farnsworth KD, Rossberg AG, Reid DG. 2012. Assessing the
 sensitivity and specificity of fish community indicators to management
 action. Can J Fish Aquat Sci 69:1065–1079

 Shin YJ, Bundy A, Shannon LJ, Simier M, Coll M, Fulton EA, Link JS, Jouffre
 D, Ojaveer H, MacKinson S, Heymans JJ, Raid T. 2010. Can simple be useful
 and reliable? Using ecological indicators to represent and compare the
 states of marine ecosystems. ICES J Mar Sci 67:717–731
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{structure and functioning indicators}
