% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMaxAge.R
\name{meanMaxAge}
\alias{meanMaxAge}
\title{Calculates the mean maximum age of fish in the community}
\usage{
meanMaxAge(X, age.table, metric = c("BIOMASS", "ABUNDANCE"), years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
designating where the observation was recorded. "SPECIES" is a numeric code
indicating the species sampled.}

\item{age.table}{table with 2 columns: "SPECIES" and "MAXAGE", the maximum
recorded age of the corresponding species.}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate indicator.}

\item{years}{vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and "MeanLifespan"
}
\description{
This function calculates the mean maximum age (MMA) of fish in
 the community for \eqn{j} areas and \eqn{i} years.
}
\details{
Mean Maximum Age (MMA): \deqn{MMA = \Sigma (age_{max,i}*B_i)/\Sigma
 B_i} where the sum is over all species \eqn{i}, and \eqn{B_i} is biomass of
 species \eqn{i}. The mean lifespan or longevity is considered to be a fixed
 parameter per species. Lifespan may vary under fishing pressure, so IndiSeas
 adopted the maximum longevity observed for each species (\eqn{age_{max,i}}).
 The variation of this indicator captures changes in species composition.

 Recommended data: Fishery independent surveys, finfish and squid.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin YJ, Bundy A, Shannon LJ, Simier M, Coll M, Fulton EA, Link JS, Jouffre
 D, Ojaveer H, MacKinson S, Heymans JJ, Raid T (2010) Can simple be useful
 and reliable? Using ecological indicators to represent and compare the
 states of marine ecosystems. ICES J Mar Sci 67:717-731
}
\seealso{
Other stability and resistance indicators: \code{\link{IVILandings}},
  \code{\link{biomassPerTL}}, \code{\link{invCVBiomass}},
  \code{\link{meanMaxL}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{stability and resistance indicators}
