% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMaxL.R
\name{meanMaxL}
\alias{meanMaxL}
\title{Calculates the mean maximum length of fish in the community}
\usage{
meanMaxL(X, table.of.length.data = "INDISEAS_MAX_LENGTH",
  metric = c("BIOMASS", "ABUNDANCE"), path = NA)
}
\arguments{
\item{X}{add text here}

\item{table.of.length.data}{add text here --or delete}

\item{metric}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
 the mean maximum length (MML) of fish in the community weighted by biomass or abundance
}
\details{
Mean Maximum Length (MML): \deqn{MML = \Sigma (L_max,i*M_i)/\Sigma
 M_i} where \eqn{L_max,i} is the maximum asymptotic length (cm) of species i,
 and \eqn{M_i} is biomass or abundance of species i (excluding
 invertebrates).

 Recommended data: Fishery independent surveys, fish.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Houle JE, Farnsworth KD, Rossberg AG, Reid DG (2012) Assessing the
 sensitivity and specificity of fish community indicators to management
 action. Can J Fish Aquat Sci 69:1065-1079

 Shin YJ, Rochet MJ, Jennings S, Field JG, Gislason H (2005) Using size-based
 indicators to evaluate the ecosystem effects of fishing. In: ICES Journal of
 Marine Science. p 384-396
}
\seealso{
Other stability and resistance indicators: \code{\link{IVILandings}},
  \code{\link{biomassPerTL}}, \code{\link{invCVBiomass}},
  \code{\link{meanMaxAge}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{stability and resistance indicators}
