% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanTLLandings.R
\name{MeanTLLandings}
\alias{MeanTLLandings}
\title{Calculates the mean trophic level of fisheries landings}
\usage{
MeanTLLandings(land = dat, cutoff = 0)
}
\arguments{
\item{land}{add text here}

\item{cutoff}{add text here}
}
\description{
This function takes a dataframe with columns **** and calculates
 the mean trophic level of fisheries landings
}
\details{
Mean trophic level of fisheries landings \eqn{TL_Land}: \deqn{TL_Land
 = \Sigma (TL_i*Y_i)/Y} where \eqn{TL_i} is the trophic level of species i,
 \eqn{Y_i} is the landings of species i, and Y is the total landings of all
 species. Trophic Level of individual species is estimated either through an
 Ecopath model or dietary analysis, or taken from a global database such as
 Fishbase.

 This indicator captures the average trophic level of the species exploited
 in the fishery. In general, this indicator reflects a gradual transition in
 landings from long-lived, high trophic level, piscivorous bottom fish toward
 short-lived, low trophic level invertebrates and planktivorous pelagic fish.

 Recommended data: Commercial fisheries landings, fish and invertebrates
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Pauly D, Christensen V, Dalsgaard J, Froese R, Torres F (1998) Fishing Down
 Marine Food Webs. Science 279:860-863
}
\seealso{
Other fishing pressure indicators: \code{\link{FishingPressure}},
  \code{\link{InverseFishingPressure}},
  \code{\link{LandByGroup}},
  \code{\link{LandingsDiversity}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fishing pressure indicators}
