% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTrophicLevelCommunity.R
\name{meanTrophicLevelCommunity}
\alias{meanTrophicLevelCommunity}
\title{Calculates the mean trophic level of the community}
\usage{
meanTrophicLevelCommunity(X, TL.table, metric = c("ABUNDANCE",
  "BIOMASS"), length.based = FALSE, years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
 "ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
 designating where the observation was recorded. "SPECIES" is a numeric code
 indicating the species sampled.

 This function can also calculate the mean trophic level of the community
 from length-based data. To do so, X should be a dataframe with seven
 columns: "YEAR", "ID", and "SPECIES", are as above. "FLEN" indicates the
 size class, and "GROUP_NAME" and "TL" are the corresponding size-based
 functional group and trophic level, respectively. "BIOMASS" is the recorded
 biomass of the size-based functional groups. Note that one SPECIES code may
 be assigned more than one GROUP_NAME if ontogentic shifts occur. For
 example, on the Scotian Shelf small (< 33 cm) and large (>33 cm) haddock
 both have SPECIES code 11. Small haddock are assigend to GROUP_NAME
 "Haddock<33" with trophic level 3.38. Large haddock are assigned to
 GROUP_NAME "Haddock33+" with trophic level 3.43.}

\item{TL.table}{dataframe with columns "SPECIES" and the corresponding "TL"
(trophic level). Set to "NULL" if length.based = TRUE.}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate indicator.}

\item{length.based}{logical variable. length.based = TRUE indicates that X is
length-based data. Default is length.based = FALSE}

\item{years}{vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and
 "MeanTLCommunity" (if length.based = FALSE) or "MeanTLCommunity_Length" if
 length.based = TRUE
}
\description{
This calculates the mean trophic level of the community weighted
 by biomass for \eqn{j} areas and \eqn{i} years.
}
\details{
Mean trophic level (TL): \deqn{TL = \Sigma TL_i*B_i)/\Sigma B_i}
 \eqn{TL_i} is trophic level of species \eqn{i}, and \eqn{B_i} is the biomass
 of species \eqn{i}.

 This indicator is based on trophic levels of all species with available
 biomass time series, weighted by annual species-specific biomass, to reflect
 the structure of the community.

 Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Christensen V. 1998. Fishery-induced changes in a marine ecosystem: insight
 frommodels of the Gulf of Thailand. J. Fish Bio. 53:128-142. Article No.
 jb980809

 Shannon L, Coll M, Bundy A, Gascuel D, Heymans JJ, Kleisner K, Lynam CP,
 Piroddi C, Tam J, Travers-Trolet M, Shin Y. 2014. Trophic level-based
 indicators to track fishing impacts across marine ecosystems. Mar. Ecol.
 Prog. Ser. 512, 115–140.
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassratio}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{predatoryFish}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{ecosystem structure and function indicators}
