% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTrophicLevelCommunity.R
\name{meanTrophicLevelCommunity}
\alias{meanTrophicLevelCommunity}
\title{Calculates the mean trophic level of the community}
\usage{
meanTrophicLevelCommunity(X, pred.data.table = "indiseas_wss_tl",
  metric = "BIOMASS")
}
\arguments{
\item{X}{add text here}

\item{metric}{set to biomass}

\item{lpred.data.table}{to delete?}
}
\description{
This function takes a dataframe with columns **** and calculates
 the mean trophic level of the community, weighted by biomass
}
\details{
Mean trophic level (TL): \deqn{TL = \Sigma TL_i*B_i)/\Sigma B_i}
 \eqn{TL_i} is trophic level of species i, and \eqn{B_i} is the biomass of
 species i.

 This indicator is based on trophic levels (TL) of all species with available
 biomass time series, weighted by annual species-specific biomass, to reflect
 the structure of the community.

 Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Christensen 1998 (not in tech report)

 Shannon et al 2014
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassRatioInv2Dem}},
  \code{\link{biomassRatioPel2Dem}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTrophicLevelCommunityLength}},
  \code{\link{predatoryFish}}, \code{\link{ratioIndicator}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{ecosystem structure and function indicators}
