% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pielouEvenness.R
\name{pielouEvenness}
\alias{pielouEvenness}
\title{Calculates Pielou's Species Evenness}
\usage{
pielouEvenness(X, group, species.table = NULL, metric = "ABUNDANCE",
  years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data or model output with
columns \code{YEAR}, \code{ID}, \code{SPECIES}, and \code{ABUNDANCE}.
\code{YEAR} indicates the year the observation was recorded, \code{ID} is an
area code indicating where the observation was recorded, \code{SPECIES} is a
numeric code indicating the species sampled, and \code{ABUNDANCE} is the
corresponding abundance (stratified and corrected for catchability as
required).}

\item{group}{A character string indicating which species to include in the
indicator calculation. If \code{group = "ALL"}, all species will be
included; otherwise, \code{group} should match a column name in
\code{species.table}.}

\item{species.table}{A table with at least one column, where the column name
is the string \code{group}, and the column entries are species codes from
\code{X} indicating which species are included that group.
\code{species.table} may also include columns for other species groups;
these will be ignored. If \code{group = "ALL"}, this table is not required.
Default is \code{species.table = NULL}.}

\item{metric}{A character string indicating which column in \code{X} to use to
calculate indicator. Default is \code{metric = "ABUNDANCE"}.}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns: \code{ID}, \code{YEAR}, and
 \code{PielouEvenness}.

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 values is assigned \code{NA}.
}
\description{
This function calculates Pielou's Species Evenness for \eqn{j}
 areas and \eqn{i} years.
}
\details{
Pielou's Species Evenness: \deqn{J' = -\Sigma p_i ln( p_i )/ln(S)}
 \eqn{p_i} is the proportion of the total sample contributed by the i(th)
 species and \eqn{S} is the number of species recorded in the sample.
 Pielou's Index is the Shannon-Weiner Index computed for the sample \eqn{S}
 and represents a measure of evenness of the community (Pielou, 1966).
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Greenstreet SP, Rogers SI. 2006. Indicators of the health of the North Sea
 fish community: identifying reference levels for an ecosystem approach to
 management. ICES J Mar Sci J du Cons 63:573 593

 Pielou EC. 1975. Ecological Diversity. Wiley, New York. ISBN 0-471-68925-4
}
\seealso{
Other biodiversity indicators: \code{\link{allBiodiversity}},
  \code{\link{heips}}, \code{\link{hillN1}},
  \code{\link{hillN2}}, \code{\link{kemptonQ}},
  \code{\link{margalef}}, \code{\link{shannon}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{biodiversity indicators}
