% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predatoryFish.R
\name{predatoryFish}
\alias{predatoryFish}
\title{Calculates the proportion of predatory fish in the community}
\usage{
predatoryFish(X, pred.data.table = "indiseas_pred_retained_wss",
  metric = "BIOMASS")
}
\arguments{
\item{X}{add text here}

\item{pred.data.table}{to delete?}

\item{metric}{set to biomass}
}
\description{
This function takes a dataframe with columns **** and calculates
 the proportion of predatory fish in the community (PropPred)
}
\details{
Proportion of predatory fish in the community is estimated by:
 \deqn{PropPred = Biomass Predatory Fish Surveyed/Total Biomass Surveyed} Predatory fish
 species are defined as all surveyed fish species that are not largely
 planktivorous (i.e. phytoplankton and zooplankton feeders should be
 excluded; Shin et al. 2010). A fish species is classified as predatory if it
 is piscivorous, or if it feeds on invertebrates that are larger than the
 macrozooplankton category (.2 cm). Detritivores should not be classified as
 predatory fish.

 This indicator captures changes in the trophic structure and changes in the
 functional diversity of fish in the ecosystem. Data used: Fishery
 independent surveys, predatory fish.

 Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin YJ, Bundy A, Shannon LJ, Simier M, Coll M, Fulton EA, Link JS, Jouffre
 D, Ojaveer H, MacKinson S, Heymans JJ, Raid T (2010) Can simple be useful
 and reliable? Using ecological indicators to represent and compare the
 states of marine ecosystems. ICES J Mar Sci 67:717-731
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassRatioInv2Dem}},
  \code{\link{biomassRatioPel2Dem}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTrophicLevelCommunityLength}},
  \code{\link{meanTrophicLevelCommunity}},
  \code{\link{ratioIndicator}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{ecosystem structure and function indicators}
