% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predatoryFish.R
\name{predatoryFish}
\alias{predatoryFish}
\title{Calculates the proportion of predatory fish in the community}
\usage{
predatoryFish(X, pred.species, metric = "BIOMASS", years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
designating where the observation was recorded. "SPECIES" is a numeric code
indicating the species sampled.}

\item{pred.species}{vector of species codes for predatory fish}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate the indicator. Default is "BIOMASS".}

\item{years}{vector of years for which to calculate indicator}
}
\value{
Returns a dataframe with 3 columns: "ID", YEAR", and
 "PropPredatoryFish". Note: if there are no observations in X for total
 biomass or predatory fish biomass in year i and spatial scale j, indicator
 value is set to NA.
}
\description{
This function calculates the proportion of predatory fish in the
 community for \eqn{j} areas and \eqn{i} years.
}
\details{
Proportion of predatory fish in the community is estimated by:
 \deqn{PropPred = Biomass Predatory Fish Surveyed/Total Biomass Surveyed}
 Predatory fish species are defined as all surveyed fish species that are not
 largely planktivorous (i.e. phytoplankton and zooplankton feeders should be
 excluded; Shin et al. 2010). A fish species is classified as predatory if it
 is piscivorous, or if it feeds on invertebrates that are larger than the
 macrozooplankton category (.2 cm). Detritivores should not be classified as
 predatory fish.

 This indicator captures changes in the trophic structure and changes in the
 functional diversity of fish in the ecosystem. Data used: Fishery
 independent surveys, predatory fish.

 Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Shin YJ, Bundy A, Shannon LJ, Simier M, Coll M, Fulton EA, Link JS, Jouffre
 D, Ojaveer H, MacKinson S, Heymans JJ, Raid T (2010) Can simple be useful
 and reliable? Using ecological indicators to represent and compare the
 states of marine ecosystems. ICES J Mar Sci 67:717-731
}
\seealso{
Other ecosystem structure and function indicators: \code{\link{biomassratio}},
  \code{\link{communityFultonK}},
  \code{\link{largeFishIndicator}},
  \code{\link{largeSpeciesIndicator}},
  \code{\link{meanTrophicLevelCommunity}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{ecosystem structure and function indicators}
