% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcePotential.R
\name{resourcePotential}
\alias{resourcePotential}
\title{Calculates the resource potential of defined species groups}
\usage{
resourcePotential(X, user.defined = F, group = c("FINFISH", "ALL",
  "SKATES", "CLUPEIDS", "GROUNDFISH", "FLATFISH", "GADOIDS",
  "INVERTEBRATES", "FORAGE", "PELAGIC", "LBENTHIVORE", "MBENTHIVORE",
  "PISCIVORE", "PLANKTIVORE", "ZOOPISCIVORE"), metric = c("BIOMASS",
  "ABUNDANCE"), herring = T, yr = 0)
}
\arguments{
\item{X}{add text here}

\item{metric}{can choose biomass or abundance}

\item{herring}{text}

\item{use.defined}{text}

\item{year}{text}
}
\description{
This function takes a dataframe with columns **** and calculates
 the biomass or abundance of a defined group of species
}
\details{
This indicator reflect temporal dynamics of seven fished groups:
 Clupeids, Finfish, Flatfish, Gadoids, Groundfish, Invertebrates and Forage
 fish (*or this might change depending on ow we tell them to organize data)

 Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\seealso{
Other resource potential indicators: \code{\link{FishingInBalance}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{resource potential indicators}
