% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcePotential.R
\name{resourcePotential}
\alias{resourcePotential}
\title{Calculates the resource potential of pre-defined species groups}
\usage{
resourcePotential(X, user.defined = F, group = c("FINFISH", "ALL",
  "SKATES", "CLUPEIDS", "GROUNDFISH", "FLATFISH", "GADOIDS",
  "INVERTEBRATES", "FORAGE", "PELAGIC", "LBENTHIVORE", "MBENTHIVORE",
  "PISCIVORE", "PLANKTIVORE", "ZOOPISCIVORE"), metric = c("BIOMASS",
  "ABUNDANCE"), years)
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
designating where the observation was recorded. "SPECIES" is a numeric code
indicating the species sampled.}

\item{user.defined}{for a user-defined group of species, put a vector of
species codes to include here. Default is "FALSE"}

\item{group}{character string indicating which group of species to include.
Note that this subsetting is based on the Fisheries and Oceans Canada
species codes for the Scotian Shelf. For other regions it may be prudent to
subsetdata to species groups of interest prior to using the function and
then choose group = "ALL". Type ?speciesgroups for more information.}

\item{metric}{character string indicating whether to use "BIOMASS" or
"ABUNDANCE" to calculate indicator.}

\item{years}{vector of years for which to calculate indicator}
}
\value{
Returns a dataframe with 3 columns. "ID", "YEAR", and "metric_group".

 If there is no data for a given year, the indicator value will be "NA" for
 that year. If biomass of species X was not captured in the survey, species
 X is still likely to be present, just not detected or perhaps not recorded.
 Replacing with zero would have an impact on trends, whereas treating as NA
 does not.
}
\description{
This function calculates the biomass or abundance of a
 pre-defined group of species for \eqn{j} areas and \eqn{i} years.
}
\details{
This indicator reflects temporal dynamics species groups.

 Recommended data: Fishery independent surveys, fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\seealso{
Other resource potential indicators: \code{\link{FishingInBalance}}
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
\concept{ecosystem structure and functioning indicators}
\concept{resource potential indicators}
