% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{Calculates Shannon's index of diversity (Magurran, 1988)}
\usage{
shannon(X, group, species.table = NULL, metric = "ABUNDANCE", years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "ABUNDANCE". "YEAR" indicates the year the observation
was recorded, "ID" is an area code indicating where the observation was
recorded, "SPECIES" is a numeric code indicating the species sampled, and
"ABUNDANCE" is the corresponding abundance (stratified and corrected for
catchability as required).}

\item{group}{A character string indicating which species to include in the
indicator calculation. If group = "ALL", all species will be included;
otherwise, group should match a column name in species.table.}

\item{species.table}{If group does not equal "ALL", species.table is a table
with at least one column, where the column name is the string group, and the
column entries are the species codes indicating the species to be included
in the calculation. species.table may also include columns for other species
groups; these will be ignored. If group = "ALL", this table is not required.
Default is species.table = NULL}

\item{metric}{A character string indicating which column in X to use to
calculate indicator. Default is "ABUNDANCE".}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns: "ID", YEAR", and
 "ShannonDiversity".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 values is assigned NA.
}
\description{
This function calculates Shannon's index of diversity for \eqn{j}
 areas and \eqn{i} years.
}
\details{
Shannon's index of diversity (H'): \deqn{H' = -\Sigma p_i ln(p_i)}
 \eqn{p_i} is the proportion of the total sample contributed by the i(th)
 species and \eqn{S} is the number of species recorded in the sample. This
 index is sensitive to the number of species recorded in the sample.

 Recommended data: Fishery independent survey data or model output; fish and
 invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.

 Greenstreet SP, Rogers SI. 2006. Indicators of the health of the North Sea
 fish community: identifying reference levels for an ecosystem approach to
 management. ICES J Mar Sci J du Cons 63:573-593

 Magurran, AE. 1988. Ecological Diversity and its Measurement. Chapman and
 Hall, London. 179 pp.
}
\seealso{
Other biodiversity indicators: \code{\link{heips}},
  \code{\link{hillN1}}, \code{\link{hillN2}},
  \code{\link{kemptonQ}}, \code{\link{margalef}},
  \code{\link{pielouEvenness}},
  \code{\link{speciesRichness}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{biodiversity indicators}
