% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesgroups.R
\name{speciesgroups}
\alias{speciesgroups}
\title{Assigns species to species groups}
\usage{
speciesgroups(X, group = c("ALL", "FINFISH", "SKATES", "CLUPEIDS",
  "GROUNDFISH", "FLATFISH", "GADOIDS", "FORAGE", "LBENTHIVORE",
  "MBENTHIVORE", "PISCIVORE", "PLANKTIVORE", "ZOOPISCIVORE"))
}
\arguments{
\item{X}{dataframe of fishery independent survey data with columns "YEAR",
"ID", "SPECIES", and "BIOMASS" and/or "ABUNDANCE". "ID" is an area code
designating where the observation was recorded. "SPECIES" is a numeric code
indicating the species sampled.}

\item{group}{character string indicating which group of species to include.
Note that this subsetting is based on the Fisheries and Oceans Canada
species codes for the Scotian Shelf. For other regions it may be prudent to
subsetdata to species groups of interest prior to using the function and
then choose group = "ALL". Type ?speciesgroups for more information.}
}
\value{
text here
}
\description{
This function assigns species into groups
 'FINFISH','SKATES','CLUPEIDS','GROUNDFISH',
 FLATFISH','GADOIDS','FORAGE','LBENTHIVORE',
 MBENTHIVORE','PISCIVORE','PLANKTIVORE', and 'ZOOPISCIVORE'
}
\details{
Will need text here explaining that specific to Scotian Shelf AND
 how to find out which species are in each group
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\author{
Danielle Dempsey, Alida Bundy, Adam Cooke, Mike McMahon,
 \email{Mike.McMahon@dfo-mpo.gc.ca}, Catalina Gomez
}
