% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesRichness.R
\name{speciesRichness}
\alias{speciesRichness}
\title{Calculates species richness of the community or the diversity of target
 species}
\usage{
speciesRichness(X, group, species.table = NULL, metric, years)
}
\arguments{
\item{X}{A dataframe of fishery independent survey data OR commercial landings
 data. Fishery independent suvey data has columns "YEAR", "ID", "SPECIES",
 and "ABUNDANCE" and/or "BIOMASS". "YEAR" indicates the year the observation
 was recorded, "ID" is an area code indicating where the observation was
 recorded, "SPECIES" is a numeric code indicating the species sampled, and
 "ABUNDANCE"/"BIOMASS" is the corresponding abundance/biomass (stratified and
 corrected for catchability as required).

 Similarly,  commercial landings data will have columns "YEAR", "ID",
 "SPECIES" are as above, and "CATCH" is the corresponding landed weight.}

\item{group}{A character string indicating which species to include in the
indicator calculation. If group = "ALL", all species will be included;
otherwise, group should match a column name in species.table.}

\item{species.table}{If group does not equal "ALL", species.table is a table
with at least one column, where the column name is the string group, and the
column entries are the species codes indicating the species from X to
include in the calculation. species.table may also include columns for other
species groups; these will be ignored. If group = "ALL", this table is not
required. Default is species.table = NULL.}

\item{metric}{A character string indicating which column in X to use to
calculate indicator. Default is "ABUNDANCE".}

\item{years}{A vector of years for which to calculate indicator.}
}
\value{
Returns a dataframe with 3 columns. If metric = "BIOMASS" or metric =
 "ABUNDANCE", columns will be named "ID", "YEAR", "SpeciesRichness". If metrc
 = "CATCH", columns will be named  "ID", "YEAR", "DiversityTargetSpp".

 If there is no data for spatial scale \eqn{j} in year \eqn{i}, indicator
 values is assigned NA.
}
\description{
This function counts the number of species recorded in fishery
 independent survey data or commercial landings data for \eqn{i} years and
 \eqn{j} areas.
}
\details{
Species richness (\eqn{S_y}) is the count of the number of species
 recorded in all research vessel trawl surveys collected year \eqn{y}.

 Recommended data: Fishery independent surveys or model output; fish and
 invertebrates

 The diversity of the target species for year y (\eqn{TS_y}) is the count of
 the number of target species recorded in all trawl catches collected in that
 year.

 Recommended data: commercial fisheries landings; fish and invertebrates.
}
\references{
Bundy A, Gomez C, Cook AM. 2017. Guidance framework for the
 selection and evaluation of ecological indicators. Can. Tech. Rep. Fish.
 Aquat. Sci. 3232: xii + 212 p.
}
\seealso{
Other biodiversity indicators: \code{\link{allBiodiversity}},
  \code{\link{heips}}, \code{\link{hillN1}},
  \code{\link{hillN2}}, \code{\link{kemptonQ}},
  \code{\link{margalef}}, \code{\link{pielouEvenness}},
  \code{\link{shannon}}

Other fishing pressure indicators: \code{\link{allPressure}},
  \code{\link{fishingPressure}}, \code{\link{landings}},
  \code{\link{meanTLLandings}}
}
\author{
Danielle Dempsey, Adam Cook \email{Adam.Cook@dfo-mpo.gc.ca},
 Catalina Gomez, Alida Bundy
}
\concept{biodiversity indicators}
\concept{fishing pressure indicators}
